@@ -11,7 +11,7 @@ import Libavcodec
 import Libavfilter
 import Libavformat
 
-final class MEPlayerItem {
+public final class MEPlayerItem: Sendable {
     private let url: URL
     private let options: KSOptions
     private let operationQueue = OperationQueue()
@@ -32,29 +32,25 @@ final class MEPlayerItem {
     private var isSeek = false
     private var allPlayerItemTracks = [PlayerItemTrackProtocol]()
     private var maxFrameDuration = 10.0
-    private var videoAudioTracks: [CapacityProtocol] {
-        var tracks = [CapacityProtocol]()
-        if let audioTrack {
-            tracks.append(audioTrack)
-        }
-        if !options.videoDisable, let videoTrack {
-            tracks.append(videoTrack)
-        }
-        return tracks
-    }
-
+    private var videoAudioTracks = [CapacityProtocol]()
+    private var audioRecognizer: AudioRecognizer?
     private var videoTrack: SyncPlayerItemTrack<VideoVTBFrame>?
     private var audioTrack: SyncPlayerItemTrack<AudioFrame>?
     private(set) var assetTracks = [FFmpegAssetTrack]()
     private var videoAdaptation: VideoAdaptationState?
-    var currentPlaybackTime: TimeInterval {
-        state == .seeking ? seekTime : mainClock().positionTime
+    private var videoDisplayCount = UInt8(0)
+    private var seekByBytes = false
+    private var lastVideoDisplayTime = CACurrentMediaTime()
+    public private(set) var chapters: [Chapter] = []
+    public var currentPlaybackTime: TimeInterval {
+        state == .seeking ? seekTime : (mainClock().time - startTime).seconds
     }
 
     private var seekTime = TimeInterval(0)
-    private(set) var startTime = TimeInterval(0)
-    private(set) var duration: TimeInterval = 0
-    private(set) var naturalSize = CGSize.zero
+    private var startTime = CMTime.zero
+    public private(set) var duration: TimeInterval = 0
+    public private(set) var fileSize: Double = 0
+    public private(set) var naturalSize = CGSize.zero
     private var error: NSError? {
         didSet {
             if error != nil {
@@ -81,13 +77,23 @@ final class MEPlayerItem {
         }
     }
 
-    private lazy var timer: Timer = .scheduledTimer(withTimeInterval: 0.02, repeats: true) { [weak self] _ in
+    private lazy var timer: Timer = .scheduledTimer(withTimeInterval: 0.05, repeats: true) { [weak self] _ in
         self?.codecDidChangeCapacity()
     }
 
+    lazy var dynamicInfo = DynamicInfo { [weak self] in
+        // metadata可能会实时变化。所以把它放在DynamicInfo里面
+        toDictionary(self?.formatCtx?.pointee.metadata)
+    } bytesRead: { [weak self] in
+        self?.formatCtx?.pointee.pb?.pointee.bytes_read ?? 0
+    } audioBitrate: { [weak self] in
+        Int(8 * (self?.audioTrack?.bitrate ?? 0))
+    } videoBitrate: { [weak self] in
+        Int(8 * (self?.videoTrack?.bitrate ?? 0))
+    }
+
     private static var onceInitial: Void = {
-        var result = SSL_library_init()
-        result = avformat_network_init()
+        var result = avformat_network_init()
         av_log_set_callback { ptr, level, format, args in
             guard let format else {
                 return
@@ -99,14 +105,15 @@ final class MEPlayerItem {
             }
             if let ptr {
                 let avclass = ptr.assumingMemoryBound(to: UnsafePointer<AVClass>.self).pointee
-                if avclass.pointee.category == AV_CLASS_CATEGORY_NA, avclass == &ffurl_context_class {
+                if avclass == &ffurl_context_class {
                     let context = ptr.assumingMemoryBound(to: URLContext.self).pointee
                     if let opaque = context.interrupt_callback.opaque {
                         let playerItem = Unmanaged<MEPlayerItem>.fromOpaque(opaque).takeUnretainedValue()
-                        playerItem.options.io(log: log)
+                        playerItem.options.urlIO(log: String(log))
                         if log.starts(with: "Will reconnect at") {
-                            playerItem.videoTrack?.seekTime = playerItem.currentPlaybackTime
-                            playerItem.audioTrack?.seekTime = playerItem.currentPlaybackTime
+                            let seconds = playerItem.mainClock().time.seconds
+                            playerItem.videoTrack?.seekTime = seconds
+                            playerItem.audioTrack?.seekTime = seconds
                         }
                     }
                 } else if avclass == avfilter_get_class() {
@@ -119,14 +126,14 @@ final class MEPlayerItem {
             }
             // 找不到解码器
             if log.hasPrefix("parser not found for codec") {
-                KSLog(log)
+                KSLog(level: .error, log)
             }
-            KSLog(level: LogLevel(rawValue: level) ?? KSOptions.logLevel, log)
+            KSLog(level: LogLevel(rawValue: level) ?? .warning, log)
         }
     }()
 
     weak var delegate: MEPlayerDelegate?
-    init(url: URL, options: KSOptions) {
+    public init(url: URL, options: KSOptions) {
         self.url = url
         self.options = options
         timer.fireDate = Date.distantFuture
@@ -134,35 +141,37 @@ final class MEPlayerItem {
         operationQueue.maxConcurrentOperationCount = 1
         operationQueue.qualityOfService = .userInteractive
         _ = MEPlayerItem.onceInitial
+        if let locale = options.audioLocale {
+            audioRecognizer = AudioRecognizer(locale: locale) { text in
+                print(text)
+            }
+        }
     }
 
-    func select(track: some MediaPlayerTrack) {
+    func select(track: some MediaPlayerTrack) -> Bool {
         if track.isEnabled {
-            return
+            return false
         }
         assetTracks.filter { $0.mediaType == track.mediaType }.forEach {
-            if $0.mediaType == .subtitle, !$0.isImageSubtitle {
-                return
-            }
-            $0.isEnabled = false
+            $0.isEnabled = track === $0
         }
-        track.isEnabled = true
         guard let assetTrack = track as? FFmpegAssetTrack else {
-            return
+            return false
         }
         if assetTrack.mediaType == .video {
             findBestAudio(videoTrack: assetTrack)
         } else if assetTrack.mediaType == .subtitle {
             if assetTrack.isImageSubtitle {
                 if !options.isSeekImageSubtitle {
-                    return
+                    return false
                 }
             } else {
-                return
+                return false
             }
         }
         seek(time: currentPlaybackTime) { _ in
         }
+        return true
     }
 }
 
@@ -198,6 +207,7 @@ extension MEPlayerItem {
 //        formatCtx.pointee.io_open = { formatCtx, context, url, flags, options -> Int32 in
 //            return 0
 //        }
+        setHttpProxy()
         var avOptions = options.formatContextOptions.avOptions
         let urlString: String
         if url.isFileURL {
@@ -224,7 +234,9 @@ extension MEPlayerItem {
         }
         options.openTime = CACurrentMediaTime()
         formatCtx.pointee.flags |= AVFMT_FLAG_GENPTS
-        av_format_inject_global_side_data(formatCtx)
+        if options.nobuffer {
+            formatCtx.pointee.flags |= AVFMT_FLAG_NOBUFFER
+        }
         if let probesize = options.probesize {
             formatCtx.pointee.probesize = probesize
         }
@@ -237,20 +249,37 @@ extension MEPlayerItem {
             avformat_close_input(&self.formatCtx)
             return
         }
-        maxFrameDuration = formatCtx.pointee.iformat.pointee.flags & AVFMT_TS_DISCONT != 0 ? 10.0 : 3600.0
+        // FIXME: hack, ffplay maybe should not use avio_feof() to test for the end
+        formatCtx.pointee.pb?.pointee.eof_reached = 0
+        let flags = formatCtx.pointee.iformat.pointee.flags
+        maxFrameDuration = flags & AVFMT_TS_DISCONT == AVFMT_TS_DISCONT ? 10.0 : 3600.0
         options.findTime = CACurrentMediaTime()
         options.formatName = String(cString: formatCtx.pointee.iformat.pointee.name)
+        seekByBytes = (flags & AVFMT_NO_BYTE_SEEK == 0) && (flags & AVFMT_TS_DISCONT != 0) && options.formatName != "ogg"
         if formatCtx.pointee.start_time != Int64.min {
-            startTime = CMTime(value: formatCtx.pointee.start_time, timescale: AV_TIME_BASE).seconds
+            startTime = CMTime(value: formatCtx.pointee.start_time, timescale: AV_TIME_BASE)
+            videoClock.time = startTime
+            audioClock.time = startTime
         }
-        mainClock().positionTime = startTime
         duration = TimeInterval(max(formatCtx.pointee.duration, 0) / Int64(AV_TIME_BASE))
-        if duration > startTime {
-            duration -= startTime
-        }
+        fileSize = Double(formatCtx.pointee.bit_rate) * duration / 8
         createCodec(formatCtx: formatCtx)
+        if formatCtx.pointee.nb_chapters > 0 {
+            chapters.removeAll()
+            for i in 0 ..< formatCtx.pointee.nb_chapters {
+                if let chapter = formatCtx.pointee.chapters[Int(i)]?.pointee {
+                    let timeBase = Timebase(chapter.time_base)
+                    let start = timeBase.cmtime(for: chapter.start).seconds
+                    let end = timeBase.cmtime(for: chapter.end).seconds
+                    let metadata = toDictionary(chapter.metadata)
+                    let title = metadata["title"] ?? ""
+                    chapters.append(Chapter(start: start, end: end, title: title))
+                }
+            }
+        }
+
         if let outputURL = options.outputURL {
-            openOutput(url: outputURL)
+            startRecord(url: outputURL)
         }
         if videoTrack == nil, audioTrack == nil {
             state = .failed
@@ -260,7 +289,8 @@ extension MEPlayerItem {
         }
     }
 
-    private func openOutput(url: URL) {
+    func startRecord(url: URL) {
+        stopRecord()
         let filename = url.isFileURL ? url.path : url.absoluteString
         var ret = avformat_alloc_output_context2(&outputFormatCtx, nil, nil, filename)
         guard let outputFormatCtx, let formatCtx else {
@@ -271,21 +301,21 @@ extension MEPlayerItem {
         var audioIndex: Int?
         var videoIndex: Int?
         let formatName = outputFormatCtx.pointee.oformat.pointee.name.flatMap { String(cString: $0) }
-        (0 ..< Int(formatCtx.pointee.nb_streams)).forEach { i in
+        for i in 0 ..< Int(formatCtx.pointee.nb_streams) {
             if let inputStream = formatCtx.pointee.streams[i] {
                 let codecType = inputStream.pointee.codecpar.pointee.codec_type
                 if [AVMEDIA_TYPE_AUDIO, AVMEDIA_TYPE_VIDEO, AVMEDIA_TYPE_SUBTITLE].contains(codecType) {
                     if codecType == AVMEDIA_TYPE_AUDIO {
                         if let audioIndex {
                             streamMapping[i] = audioIndex
-                            return
+                            continue
                         } else {
                             audioIndex = index
                         }
                     } else if codecType == AVMEDIA_TYPE_VIDEO {
                         if let videoIndex {
                             streamMapping[i] = videoIndex
-                            return
+                            continue
                         } else {
                             videoIndex = index
                         }
@@ -322,31 +352,27 @@ extension MEPlayerItem {
         videoAdaptation = nil
         videoTrack = nil
         audioTrack = nil
+        videoAudioTracks.removeAll()
         assetTracks = (0 ..< Int(formatCtx.pointee.nb_streams)).compactMap { i in
             if let coreStream = formatCtx.pointee.streams[i] {
                 coreStream.pointee.discard = AVDISCARD_ALL
                 if let assetTrack = FFmpegAssetTrack(stream: coreStream) {
-                    assetTrack.startTime = startTime
-                    if !options.subtitleDisable, assetTrack.mediaType == .subtitle {
-                        let subtitle = SyncPlayerItemTrack<SubtitleFrame>(assetTrack: assetTrack, options: options)
-                        assetTrack.isEnabled = !assetTrack.isImageSubtitle
+                    if assetTrack.mediaType == .subtitle {
+                        let subtitle = SyncPlayerItemTrack<SubtitleFrame>(mediaType: .subtitle, frameCapacity: 255, options: options)
                         assetTrack.subtitle = subtitle
                         allPlayerItemTracks.append(subtitle)
                     }
+                    assetTrack.seekByBytes = seekByBytes
                     return assetTrack
                 }
             }
             return nil
         }
-        if options.autoSelectEmbedSubtitle {
-            assetTracks.first { $0.mediaType == .subtitle }?.isEnabled = true
-        }
         var videoIndex: Int32 = -1
         if !options.videoDisable {
             let videos = assetTracks.filter { $0.mediaType == .video }
-            let bitRates = videos.map(\.bitRate)
             let wantedStreamNb: Int32
-            if videos.count > 0, let index = options.wantedVideo(bitRates: bitRates) {
+            if !videos.isEmpty, let index = options.wantedVideo(tracks: videos) {
                 wantedStreamNb = videos[index].trackID
             } else {
                 wantedStreamNb = -1
@@ -369,11 +395,19 @@ extension MEPlayerItem {
                     }
                 }
                 naturalSize = abs(rotation - 90) <= 1 || abs(rotation - 270) <= 1 ? first.naturalSize.reverse : first.naturalSize
-                let track = options.syncDecodeVideo ? SyncPlayerItemTrack<VideoVTBFrame>(assetTrack: first, options: options) : AsyncPlayerItemTrack<VideoVTBFrame>(assetTrack: first, options: options)
+                options.process(assetTrack: first)
+                let frameCapacity = options.videoFrameMaxCount(fps: first.nominalFrameRate, naturalSize: naturalSize, isLive: duration == 0)
+                let track = options.syncDecodeVideo ? SyncPlayerItemTrack<VideoVTBFrame>(mediaType: .video, frameCapacity: frameCapacity, options: options) : AsyncPlayerItemTrack<VideoVTBFrame>(mediaType: .video, frameCapacity: frameCapacity, options: options)
                 track.delegate = self
                 allPlayerItemTracks.append(track)
                 videoTrack = track
-                if videos.count > 1, options.videoAdaptable {
+                if first.codecpar.codec_id != AV_CODEC_ID_MJPEG {
+                    videoAudioTracks.append(track)
+                }
+                let bitRates = videos.map(\.bitRate).filter {
+                    $0 > 0
+                }
+                if bitRates.count > 1, options.videoAdaptable {
                     let bitRateState = VideoAdaptationState.BitRateState(bitRate: first.bitRate, time: CACurrentMediaTime())
                     videoAdaptation = VideoAdaptationState(bitRates: bitRates.sorted(by: <), duration: duration, fps: first.nominalFrameRate, bitRateStates: [bitRateState])
                 }
@@ -382,7 +416,7 @@ extension MEPlayerItem {
 
         let audios = assetTracks.filter { $0.mediaType == .audio }
         let wantedStreamNb: Int32
-        if audios.count > 0, let index = options.wantedAudio(infos: audios.map { ($0.bitRate, $0.language) }) {
+        if !audios.isEmpty, let index = options.wantedAudio(tracks: audios) {
             wantedStreamNb = audios[index].trackID
         } else {
             wantedStreamNb = -1
@@ -390,12 +424,17 @@ extension MEPlayerItem {
         let index = av_find_best_stream(formatCtx, AVMEDIA_TYPE_AUDIO, wantedStreamNb, videoIndex, nil, 0)
         if let first = audios.first(where: {
             index > 0 ? $0.trackID == index : true
-        }) {
+        }), first.codecpar.codec_id != AV_CODEC_ID_NONE {
             first.isEnabled = true
-            let track = options.syncDecodeAudio ? SyncPlayerItemTrack<AudioFrame>(assetTrack: first, options: options) : AsyncPlayerItemTrack<AudioFrame>(assetTrack: first, options: options)
+            options.process(assetTrack: first)
+            // 音频要比较所有的音轨，因为truehd的fps是1200，跟其他的音轨差距太大了
+            let fps = audios.map(\.nominalFrameRate).max() ?? 44
+            let frameCapacity = options.audioFrameMaxCount(fps: fps, channelCount: Int(first.audioDescriptor?.audioFormat.channelCount ?? 2))
+            let track = options.syncDecodeAudio ? SyncPlayerItemTrack<AudioFrame>(mediaType: .audio, frameCapacity: frameCapacity, options: options) : AsyncPlayerItemTrack<AudioFrame>(mediaType: .audio, frameCapacity: frameCapacity, options: options)
             track.delegate = self
             allPlayerItemTracks.append(track)
             audioTrack = track
+            videoAudioTracks.append(track)
             isAudioStalled = false
         }
     }
@@ -417,11 +456,15 @@ extension MEPlayerItem {
     private func readThread() {
         if state == .opened {
             if options.startPlayTime > 0 {
-                seekTime = options.startPlayTime + startTime
-                state = .seeking
-            } else {
-                state = .reading
+                let timestamp = startTime + CMTime(seconds: options.startPlayTime)
+                let flags = seekByBytes ? AVSEEK_FLAG_BYTE : 0
+                let seekStartTime = CACurrentMediaTime()
+                let result = avformat_seek_file(formatCtx, -1, Int64.min, timestamp.value, Int64.max, flags)
+                audioClock.time = timestamp
+                videoClock.time = timestamp
+                KSLog("start PlayTime: \(timestamp.seconds) spend Time: \(CACurrentMediaTime() - seekStartTime)")
             }
+            state = .reading
         }
         allPlayerItemTracks.forEach { $0.decode() }
         while [MESourceState.paused, .seeking, .reading].contains(state) {
@@ -429,76 +472,103 @@ extension MEPlayerItem {
                 condition.wait()
             }
             if state == .seeking {
-                let time = seekTime
-                let timeStamp = Int64(time * TimeInterval(AV_TIME_BASE))
-                let startTime = CACurrentMediaTime()
+                let seekToTime = seekTime
+                let time = mainClock().time
+                var increase = Int64(seekTime + startTime.seconds - time.seconds)
+                var seekFlags = options.seekFlags
+                let timeStamp: Int64
+                if seekByBytes {
+                    seekFlags |= AVSEEK_FLAG_BYTE
+                    if let bitRate = formatCtx?.pointee.bit_rate {
+                        increase = increase * bitRate / 8
+                    } else {
+                        increase *= 180_000
+                    }
+                    var position = Int64(-1)
+                    if position < 0 {
+                        position = videoClock.position
+                    }
+                    if position < 0 {
+                        position = audioClock.position
+                    }
+                    if position < 0 {
+                        position = avio_tell(formatCtx?.pointee.pb)
+                    }
+                    timeStamp = position + increase
+                } else {
+                    increase *= Int64(AV_TIME_BASE)
+                    timeStamp = Int64(time.seconds) * Int64(AV_TIME_BASE) + increase
+                }
+                let seekMin = increase > 0 ? timeStamp - increase + 2 : Int64.min
+                let seekMax = increase < 0 ? timeStamp - increase - 2 : Int64.max
                 // can not seek to key frame
-//                let result = avformat_seek_file(formatCtx, -1, Int64.min, timeStamp, Int64.max, options.seekFlags)
-                var result = av_seek_frame(formatCtx, -1, timeStamp, options.seekFlags)
+                let seekStartTime = CACurrentMediaTime()
+                var result = avformat_seek_file(formatCtx, -1, seekMin, timeStamp, seekMax, seekFlags)
+//                var result = av_seek_frame(formatCtx, -1, timeStamp, seekFlags)
                 // When seeking before the beginning of the file, and seeking fails,
                 // try again without the backwards flag to make it seek to the
                 // beginning.
-                if result < 0, options.seekFlags & AVSEEK_FLAG_BACKWARD > 0 {
+                if result < 0, seekFlags & AVSEEK_FLAG_BACKWARD == AVSEEK_FLAG_BACKWARD {
+                    KSLog("seek to \(seekToTime) failed. seekFlags remove BACKWARD")
                     options.seekFlags &= ~AVSEEK_FLAG_BACKWARD
-                    result = av_seek_frame(formatCtx, -1, timeStamp, options.seekFlags)
+                    seekFlags &= ~AVSEEK_FLAG_BACKWARD
+                    result = avformat_seek_file(formatCtx, -1, seekMin, timeStamp, seekMax, seekFlags)
                 }
-                KSLog("seek to \(time) spend Time: \(CACurrentMediaTime() - startTime)")
+                KSLog("seek to \(seekToTime) spend Time: \(CACurrentMediaTime() - seekStartTime)")
                 if state == .closed {
                     break
                 }
+                if seekToTime != seekTime {
+                    continue
+                }
                 isSeek = true
-                allPlayerItemTracks.forEach { $0.seek(time: time) }
+                allPlayerItemTracks.forEach { $0.seek(time: seekToTime) }
                 DispatchQueue.main.async { [weak self] in
                     guard let self else { return }
                     self.seekingCompletionHandler?(result >= 0)
                     self.seekingCompletionHandler = nil
                 }
-                mainClock().positionTime = time
-                videoClock.positionTime = time
-                videoClock.duration = 0
+                audioClock.time = CMTime(seconds: seekToTime, preferredTimescale: time.timescale) + startTime
+                videoClock.time = CMTime(seconds: seekToTime, preferredTimescale: time.timescale) + startTime
                 state = .reading
-
             } else if state == .reading {
                 autoreleasepool {
-                    reading()
+                    _ = reading()
                 }
             }
         }
     }
 
-    private func reading() {
+    private func reading() -> Int32 {
         let packet = Packet()
         guard let corePacket = packet.corePacket else {
-            return
+            return 0
         }
         let readResult = av_read_frame(formatCtx, corePacket)
         if state == .closed {
-            return
+            return 0
         }
         if readResult == 0 {
             if let outputFormatCtx, let formatCtx {
                 let index = Int(corePacket.pointee.stream_index)
                 if let outputIndex = streamMapping[index],
                    let inputTb = formatCtx.pointee.streams[index]?.pointee.time_base,
-                   let outputTb = outputFormatCtx.pointee.streams[outputIndex]?.pointee.time_base
+                   let outputTb = outputFormatCtx.pointee.streams[outputIndex]?.pointee.time_base,
+                   let outputPacket
                 {
                     av_packet_ref(outputPacket, corePacket)
-                    outputPacket?.pointee.stream_index = Int32(outputIndex)
+                    outputPacket.pointee.stream_index = Int32(outputIndex)
                     av_packet_rescale_ts(outputPacket, inputTb, outputTb)
-                    outputPacket?.pointee.pos = -1
+                    outputPacket.pointee.pos = -1
                     let ret = av_interleaved_write_frame(outputFormatCtx, outputPacket)
                     if ret < 0 {
                         KSLog("can not av_interleaved_write_frame")
                     }
                 }
             }
-            if formatCtx?.pointee.pb?.pointee.eof_reached == 1 {
-                // todo need reconnect
-            }
             if corePacket.pointee.size <= 0 {
-                return
+                return 0
             }
-            packet.fill()
             let first = assetTracks.first { $0.trackID == corePacket.pointee.stream_index }
             if let first, first.isEnabled {
                 packet.assetTrack = first
@@ -517,10 +587,10 @@ extension MEPlayerItem {
                 }
             }
         } else {
-            if readResult == AVError.eof.code || formatCtx?.pointee.pb?.pointee.eof_reached == 1 {
+            if readResult == AVError.eof.code || avio_feof(formatCtx?.pointee.pb) > 0 {
                 if options.isLoopPlay, allPlayerItemTracks.allSatisfy({ !$0.isLoopModel }) {
                     allPlayerItemTracks.forEach { $0.isLoopModel = true }
-                    _ = av_seek_frame(formatCtx, -1, Int64(startTime), AVSEEK_FLAG_BACKWARD)
+                    _ = av_seek_frame(formatCtx, -1, startTime.value, AVSEEK_FLAG_BACKWARD)
                 } else {
                     allPlayerItemTracks.forEach { $0.isEndOfFile = true }
                     state = .finished
@@ -530,6 +600,7 @@ extension MEPlayerItem {
                 error = .init(errorCode: .readFrame, avErrorCode: readResult)
             }
         }
+        return readResult
     }
 
     private func pause() {
@@ -543,16 +614,6 @@ extension MEPlayerItem {
             state = .reading
             condition.signal()
         }
-    }
-}
-
-extension MEPlayerItem {
-    var metadata: [String: String] {
-        toDictionary(formatCtx?.pointee.metadata)
-    }
-
-    var bytesRead: Int64 {
-        formatCtx?.pointee.pb.pointee.bytes_read ?? 0
     }
 }
 
@@ -570,7 +631,7 @@ extension MEPlayerItem: MediaPlayback {
         return seekable && duration > 0
     }
 
-    func prepareToPlay() {
+    public func prepareToPlay() {
         state = .opening
         openOperation = BlockOperation { [weak self] in
             guard let self else { return }
@@ -585,13 +646,11 @@ extension MEPlayerItem: MediaPlayback {
         }
     }
 
-    func shutdown() {
+    public func shutdown() {
         guard state != .closed else { return }
         state = .closed
         av_packet_free(&outputPacket)
-        if let outputFormatCtx {
-            av_write_trailer(outputFormatCtx)
-        }
+        stopRecord()
         // 故意循环引用。等结束了。才释放
         let closeOperation = BlockOperation {
             Thread.current.name = (self.operationQueue.name ?? "") + "_close"
@@ -624,7 +683,13 @@ extension MEPlayerItem: MediaPlayback {
         self.closeOperation = closeOperation
     }
 
-    func seek(time: TimeInterval, completion: @escaping ((Bool) -> Void)) {
+    func stopRecord() {
+        if let outputFormatCtx {
+            av_write_trailer(outputFormatCtx)
+        }
+    }
+
+    public func seek(time: TimeInterval, completion: @escaping ((Bool) -> Void)) {
         if state == .reading || state == .paused {
             seekTime = time
             state = .seeking
@@ -636,6 +701,9 @@ extension MEPlayerItem: MediaPlayback {
             state = .seeking
             seekingCompletionHandler = completion
             read()
+        } else if state == .seeking {
+            seekTime = time
+            seekingCompletionHandler = completion
         }
         isAudioStalled = audioTrack == nil
     }
@@ -673,7 +741,7 @@ extension MEPlayerItem: CodecCapacityDelegate {
                 audioTrack?.isLoopModel = false
                 videoTrack?.isLoopModel = false
                 if state == .finished {
-                    seek(time: startTime) { _ in }
+                    seek(time: 0) { _ in }
                 }
             }
         }
@@ -719,39 +787,83 @@ extension MEPlayerItem: OutputRenderSourceDelegate {
         isAudioStalled ? videoClock : audioClock
     }
 
-    func setVideo(time: CMTime, duration: CMTime) {
-        videoClock.positionTime = time.seconds
-        videoClock.duration = duration.seconds
+    public func setVideo(time: CMTime, position: Int64) {
+//        print("[video] video interval \(CACurrentMediaTime() - videoClock.lastMediaTime) video diff \(time.seconds - videoClock.time.seconds)")
+        videoClock.time = time
+        videoClock.position = position
+        videoDisplayCount += 1
+        let diff = videoClock.lastMediaTime - lastVideoDisplayTime
+        if diff > 1 {
+            dynamicInfo.displayFPS = Double(videoDisplayCount) / diff
+            videoDisplayCount = 0
+            lastVideoDisplayTime = videoClock.lastMediaTime
+        }
     }
 
-    func setAudio(time: CMTime) {
-        audioClock.positionTime = time.seconds
+    public func setAudio(time: CMTime, position: Int64) {
+//        print("[audio] setAudio: \(time.seconds)")
+        audioClock.time = time
+        audioClock.position = position
     }
 
-    func getVideoOutputRender(force: Bool) -> VideoVTBFrame? {
-        guard let videoTrack, videoTrack.frameCount > 0 else {
+    public func getVideoOutputRender(force: Bool) -> VideoVTBFrame? {
+        guard let videoTrack else {
             return nil
         }
-        let type = force ? .next : options.videoClockSync(main: mainClock(), video: videoClock)
+        var type: ClockProcessType = force ? .next : .remain
+        let predicate: ((VideoVTBFrame, Int) -> Bool)? = force ? nil : { [weak self] frame, count -> Bool in
+            guard let self else { return true }
+            (self.dynamicInfo.audioVideoSyncDiff, type) = self.options.videoClockSync(main: self.mainClock(), nextVideoTime: frame.seconds, fps: frame.fps, frameCount: count)
+            return type != .remain
+        }
+        let frame = videoTrack.getOutputRender(where: predicate)
         switch type {
         case .remain:
-            return nil
+            break
         case .next:
-            return videoTrack.getOutputRender(where: nil)
-        case .dropNext:
-            _ = videoTrack.getOutputRender(where: nil)
-            return videoTrack.getOutputRender(where: nil)
+            break
+        case .dropNextFrame:
+            if videoTrack.getOutputRender(where: nil) != nil {
+                dynamicInfo.droppedVideoFrameCount += 1
+            }
         case .flush:
+            let count = videoTrack.outputRenderQueue.count
             videoTrack.outputRenderQueue.flush()
-            return nil
+            dynamicInfo.droppedVideoFrameCount += UInt32(count)
         case .seek:
             videoTrack.outputRenderQueue.flush()
-            videoTrack.seekTime = currentPlaybackTime
-            return nil
+            videoTrack.seekTime = mainClock().time.seconds
+        case .dropNextPacket:
+            if let videoTrack = videoTrack as? AsyncPlayerItemTrack {
+                let packet = videoTrack.packetQueue.pop { item, _ -> Bool in
+                    !item.isKeyFrame
+                }
+                if packet != nil {
+                    dynamicInfo.droppedVideoPacketCount += 1
+                }
+            }
+        case .dropGOPPacket:
+            if let videoTrack = videoTrack as? AsyncPlayerItemTrack {
+                var packet: Packet? = nil
+                repeat {
+                    packet = videoTrack.packetQueue.pop { item, _ -> Bool in
+                        !item.isKeyFrame
+                    }
+                    if packet != nil {
+                        dynamicInfo.droppedVideoPacketCount += 1
+                    }
+                } while packet != nil
+            }
         }
+        return frame
     }
 
-    func getAudioOutputRender() -> AudioFrame? {
-        audioTrack?.getOutputRender(where: nil)
+    public func getAudioOutputRender() -> AudioFrame? {
+        if let frame = audioTrack?.getOutputRender(where: nil) {
+            audioRecognizer?.append(frame: frame)
+            return frame
+        } else {
+            return nil
+        }
     }
 }
