@@ -5,35 +5,47 @@
 //  Created by kintan on 2018/8/7.
 //
 import Foundation
+
 public class EmptySubtitleInfo: SubtitleInfo {
+    public var isEnabled: Bool = true
     public let subtitleID: String = ""
     public var delay: TimeInterval = 0
     public let name = NSLocalizedString("no show subtitle", comment: "")
-    public func search(for _: TimeInterval) -> SubtitlePart? {
-        nil
+    public func search(for _: TimeInterval) -> [SubtitlePart] {
+        []
     }
-
-    public func subtitle(isEnabled _: Bool) {}
 }
 
-public class URLSubtitleInfo: KSURLSubtitle, SubtitleInfo {
-    private var downloadURL: URL
+public class URLSubtitleInfo: KSSubtitle, SubtitleInfo {
+    public var isEnabled: Bool = false {
+        didSet {
+            if isEnabled, parts.isEmpty {
+                Task {
+                    try? await parse(url: downloadURL, userAgent: userAgent)
+                }
+            }
+        }
+    }
+
+    public private(set) var downloadURL: URL
     public var delay: TimeInterval = 0
     public private(set) var name: String
     public let subtitleID: String
     public var comment: String?
     public var userInfo: NSMutableDictionary?
+    private let userAgent: String?
     public convenience init(url: URL) {
         self.init(subtitleID: url.absoluteString, name: url.lastPathComponent, url: url)
     }
 
-    public init(subtitleID: String, name: String, url: URL) {
+    public init(subtitleID: String, name: String, url: URL, userAgent: String? = nil) {
         self.subtitleID = subtitleID
         self.name = name
+        self.userAgent = userAgent
         downloadURL = url
         super.init()
-        if !url.isFileURL, name.count == 0 {
-            url.download { [weak self] filename, tmpUrl in
+        if !url.isFileURL, name.isEmpty {
+            url.download(userAgent: userAgent) { [weak self] filename, tmpUrl in
                 guard let self else {
                     return
                 }
@@ -46,122 +58,130 @@ public class URLSubtitleInfo: KSURLSubtitle, SubtitleInfo {
             }
         }
     }
-
-    public func subtitle(isEnabled: Bool) {
-        if isEnabled, parts.count == 0 {
-            Task {
-                try? await parse(url: downloadURL)
-            }
-        }
-    }
 }
 
 public protocol SubtitleDataSouce: AnyObject {
     var infos: [any SubtitleInfo] { get }
 }
 
+public protocol FileURLSubtitleDataSouce: SubtitleDataSouce {
+    func searchSubtitle(fileURL: URL?) async throws
+}
+
+public protocol CacheSubtitleDataSouce: FileURLSubtitleDataSouce {
+    func addCache(fileURL: URL, downloadURL: URL)
+}
+
 public protocol SearchSubtitleDataSouce: SubtitleDataSouce {
-    func searchSubtitle(url: URL) async throws
+    func searchSubtitle(query: String?, languages: [String]) async throws
 }
 
 public extension KSOptions {
     static var subtitleDataSouces: [SubtitleDataSouce] = [DirectorySubtitleDataSouce()]
 }
 
-public extension SubtitleDataSouce {
-    func addCache(subtitleID: String, downloadURL: URL) {
-        CacheDataSouce.singleton.addCache(subtitleID: subtitleID, downloadURL: downloadURL)
-    }
-}
-
-public class CacheDataSouce: SearchSubtitleDataSouce {
-    public static let singleton = CacheDataSouce()
+public class PlistCacheSubtitleDataSouce: CacheSubtitleDataSouce {
+    public static let singleton = PlistCacheSubtitleDataSouce()
     public var infos = [any SubtitleInfo]()
-    private let cacheFolder = (NSTemporaryDirectory() as NSString).appendingPathComponent("KSSubtitleCache")
-    private var srtCacheInfoPath: String
+    private let srtCacheInfoPath: String
     // 因为plist不能保存URL
-    private var srtInfoCaches = [String: String]()
+    private var srtInfoCaches: [String: [String]]
     private init() {
+        let cacheFolder = (NSTemporaryDirectory() as NSString).appendingPathComponent("KSSubtitleCache")
         if !FileManager.default.fileExists(atPath: cacheFolder) {
             try? FileManager.default.createDirectory(atPath: cacheFolder, withIntermediateDirectories: true, attributes: nil)
         }
         srtCacheInfoPath = (cacheFolder as NSString).appendingPathComponent("KSSrtInfo.plist")
-    }
-
-    public func searchSubtitle(url: URL) async throws {
-        infos.removeAll()
-        srtCacheInfoPath = (cacheFolder as NSString).appendingPathComponent("KSSrtInfo_\(url.lastPathComponent).plist")
-        if FileManager.default.fileExists(atPath: srtCacheInfoPath), let files = NSMutableDictionary(contentsOfFile: srtCacheInfoPath) as? [String: String] {
-            srtInfoCaches = files.filter { FileManager.default.fileExists(atPath: $1) }
-            if srtInfoCaches.isEmpty {
-                infos = []
-            } else {
-                let array = srtInfoCaches.map { subtitleID, downloadURL -> (any SubtitleInfo) in
-                    let info = URLSubtitleInfo(subtitleID: subtitleID, name: (downloadURL as NSString).lastPathComponent, url: URL(fileURLWithPath: downloadURL))
-                    info.comment = "本地"
-                    return info
-                }
-                infos = array
+        srtInfoCaches = [String: [String]]()
+        DispatchQueue.global().async { [weak self] in
+            guard let self else {
+                return
             }
-            if srtInfoCaches.count != files.count {
-                (srtInfoCaches as NSDictionary).write(toFile: srtCacheInfoPath, atomically: false)
-            }
-        } else {
-            srtInfoCaches = [String: String]()
-            infos = []
+            self.srtInfoCaches = (NSMutableDictionary(contentsOfFile: self.srtCacheInfoPath) as? [String: [String]]) ?? [String: [String]]()
         }
     }
 
-    public func addCache(subtitleID: String, downloadURL: URL) {
-        srtInfoCaches[subtitleID] = downloadURL.path
-        (srtInfoCaches as NSDictionary).write(toFile: srtCacheInfoPath, atomically: false)
+    public func searchSubtitle(fileURL: URL?) async throws {
+        infos = [any SubtitleInfo]()
+        guard let fileURL else {
+            return
+        }
+        infos = srtInfoCaches[fileURL.absoluteString]?.compactMap { downloadURL -> (any SubtitleInfo)? in
+            guard let url = URL(string: downloadURL) else {
+                return nil
+            }
+            let info = URLSubtitleInfo(url: url)
+            info.comment = "local"
+            return info
+        } ?? [any SubtitleInfo]()
+    }
+
+    public func addCache(fileURL: URL, downloadURL: URL) {
+        let file = fileURL.absoluteString
+        let path = downloadURL.absoluteString
+        var array = srtInfoCaches[file] ?? [String]()
+        if !array.contains(where: { $0 == path }) {
+            array.append(path)
+            srtInfoCaches[file] = array
+            DispatchQueue.global().async { [weak self] in
+                guard let self else {
+                    return
+                }
+                (self.srtInfoCaches as NSDictionary).write(toFile: self.srtCacheInfoPath, atomically: false)
+            }
+        }
     }
 }
 
 public class URLSubtitleDataSouce: SubtitleDataSouce {
-    public let infos: [any SubtitleInfo]
+    public var infos: [any SubtitleInfo]
     public init(urls: [URL]) {
         infos = urls.map { URLSubtitleInfo(url: $0) }
     }
 }
 
-public class DirectorySubtitleDataSouce: SearchSubtitleDataSouce {
+public class DirectorySubtitleDataSouce: FileURLSubtitleDataSouce {
     public var infos = [any SubtitleInfo]()
     public init() {}
 
-    public func searchSubtitle(url: URL) async throws {
-        infos.removeAll()
-        if url.isFileURL {
-            let subtitleURLs: [URL] = (try? FileManager.default.contentsOfDirectory(at: url.deletingLastPathComponent(), includingPropertiesForKeys: nil).filter(\.isSubtitle)) ?? []
-            let contents = subtitleURLs.map { URLSubtitleInfo(url: $0) }.sorted { left, right in
+    public func searchSubtitle(fileURL: URL?) async throws {
+        infos = [any SubtitleInfo]()
+        guard let fileURL else {
+            return
+        }
+        if fileURL.isFileURL {
+            let subtitleURLs: [URL] = (try? FileManager.default.contentsOfDirectory(at: fileURL.deletingLastPathComponent(), includingPropertiesForKeys: nil).filter(\.isSubtitle)) ?? []
+            infos = subtitleURLs.map { URLSubtitleInfo(url: $0) }.sorted { left, right in
                 left.name < right.name
             }
-            infos.append(contentsOf: contents)
         }
     }
 }
 
-public class ShooterSubtitleDataSouce: SearchSubtitleDataSouce {
+public class ShooterSubtitleDataSouce: FileURLSubtitleDataSouce {
     public var infos = [any SubtitleInfo]()
     public init() {}
-    public func searchSubtitle(url: URL) async throws {
-        infos.removeAll()
-        guard url.isFileURL, let url = URL(string: "https://www.shooter.cn/api/subapi.php")?
-            .add(queryItems: ["format": "json", "pathinfo": url.path, "filehash": url.shooterFilehash])
+    public func searchSubtitle(fileURL: URL?) async throws {
+        infos = [any SubtitleInfo]()
+        guard let fileURL else {
+            return
+        }
+        guard fileURL.isFileURL, let searchApi = URL(string: "https://www.shooter.cn/api/subapi.php")?
+            .add(queryItems: ["format": "json", "pathinfo": fileURL.path, "filehash": fileURL.shooterFilehash])
         else {
             return
         }
-        var request = URLRequest(url: url)
+        var request = URLRequest(url: searchApi)
         request.httpMethod = "POST"
         let (data, _) = try await URLSession.shared.data(for: request)
         guard let json = try JSONSerialization.jsonObject(with: data) as? [[String: Any]] else {
             return
         }
-        json.forEach { sub in
+        infos = json.flatMap { sub in
             let filesDic = sub["Files"] as? [[String: String]]
 //                let desc = sub["Desc"] as? String ?? ""
             let delay = TimeInterval(sub["Delay"] as? Int ?? 0) / 1000.0
-            let result = filesDic?.compactMap { dic in
+            return filesDic?.compactMap { dic in
                 if let string = dic["Link"], let url = URL(string: string) {
                     let info = URLSubtitleInfo(subtitleID: string, name: "", url: url)
                     info.delay = delay
@@ -169,7 +189,6 @@ public class ShooterSubtitleDataSouce: SearchSubtitleDataSouce {
                 }
                 return nil
             } ?? [URLSubtitleInfo]()
-            self.infos.append(contentsOf: result)
         }
     }
 }
@@ -181,9 +200,11 @@ public class AssrtSubtitleDataSouce: SearchSubtitleDataSouce {
         self.token = token
     }
 
-    public func searchSubtitle(url: URL) async throws {
-        infos.removeAll()
-        let query = url.deletingPathExtension().lastPathComponent
+    public func searchSubtitle(query: String?, languages _: [String] = ["zh-cn"]) async throws {
+        infos = [any SubtitleInfo]()
+        guard let query else {
+            return
+        }
         guard let searchApi = URL(string: "https://api.assrt.net/v1/sub/search")?.add(queryItems: ["q": query]) else {
             return
         }
@@ -200,11 +221,13 @@ public class AssrtSubtitleDataSouce: SearchSubtitleDataSouce {
         guard let subDict = json["sub"] as? [String: Any], let subArray = subDict["subs"] as? [[String: Any]] else {
             return
         }
+        var result = [URLSubtitleInfo]()
         for sub in subArray {
             if let assrtSubID = sub["id"] as? Int {
-                try await infos.append(contentsOf: loadDetails(assrtSubID: String(assrtSubID)))
+                try await result.append(contentsOf: loadDetails(assrtSubID: String(assrtSubID)))
             }
         }
+        infos = result
     }
 
     func loadDetails(assrtSubID: String) async throws -> [URLSubtitleInfo] {
@@ -226,7 +249,7 @@ public class AssrtSubtitleDataSouce: SearchSubtitleDataSouce {
             return infos
         }
         if let fileList = sub["filelist"] as? [[String: String]] {
-            fileList.forEach { dic in
+            for dic in fileList {
                 if let urlString = dic["url"], let filename = dic["f"], let url = URL(string: urlString) {
                     let info = URLSubtitleInfo(subtitleID: urlString, name: filename, url: url)
                     infos.append(info)
@@ -237,6 +260,98 @@ public class AssrtSubtitleDataSouce: SearchSubtitleDataSouce {
             infos.append(info)
         }
         return infos
+    }
+}
+
+public class OpenSubtitleDataSouce: SearchSubtitleDataSouce {
+    private var token: String? = nil
+    private let username: String?
+    private let password: String?
+    private let apiKey: String
+    public var infos = [any SubtitleInfo]()
+    public init(apiKey: String, username: String? = nil, password: String? = nil) {
+        self.apiKey = apiKey
+        self.username = username
+        self.password = password
+    }
+
+    public func searchSubtitle(query: String?, languages: [String] = ["zh-cn"]) async throws {
+        try await searchSubtitle(query: query, imdbID: 0, tmdbID: 0, languages: languages)
+    }
+
+    public func searchSubtitle(query: String?, imdbID: Int, tmdbID: Int, languages: [String] = ["zh-cn"]) async throws {
+        infos = [any SubtitleInfo]()
+        var queryItems = [String: String]()
+        if let query {
+            queryItems["query"] = query
+        }
+        if imdbID != 0 {
+            queryItems["imbd_id"] = String(imdbID)
+        }
+        if tmdbID != 0 {
+            queryItems["tmdb_id"] = String(tmdbID)
+        }
+        if queryItems.isEmpty {
+            return
+        }
+        queryItems["languages"] = languages.joined(separator: ",")
+        try await searchSubtitle(queryItems: queryItems)
+    }
+
+    // https://opensubtitles.stoplight.io/docs/opensubtitles-api/a172317bd5ccc-search-for-subtitles
+    public func searchSubtitle(queryItems: [String: String]) async throws {
+        infos = [any SubtitleInfo]()
+        if queryItems.isEmpty {
+            return
+        }
+        guard let searchApi = URL(string: "https://api.opensubtitles.com/api/v1/subtitles")?.add(queryItems: queryItems) else {
+            return
+        }
+        var request = URLRequest(url: searchApi)
+        request.addValue(apiKey, forHTTPHeaderField: "Api-Key")
+        if let token {
+            request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
+        }
+        let (data, _) = try await URLSession.shared.data(for: request)
+        guard let json = try JSONSerialization.jsonObject(with: data) as? [String: Any] else {
+            return
+        }
+        guard let dataArray = json["data"] as? [[String: Any]] else {
+            return
+        }
+        var result = [URLSubtitleInfo]()
+        for sub in dataArray {
+            if let attributes = sub["attributes"] as? [String: Any], let files = attributes["files"] as? [[String: Any]] {
+                for file in files {
+                    if let fileID = file["file_id"] as? Int, let info = try await loadDetails(fileID: fileID) {
+                        result.append(info)
+                    }
+                }
+            }
+        }
+        infos = result
+    }
+
+    func loadDetails(fileID: Int) async throws -> URLSubtitleInfo? {
+        guard let detailApi = URL(string: "https://api.opensubtitles.com/api/v1/download")?.add(queryItems: ["file_id": String(fileID)]) else {
+            return nil
+        }
+        var request = URLRequest(url: detailApi)
+        request.httpMethod = "POST"
+        request.addValue(apiKey, forHTTPHeaderField: "Api-Key")
+        if let token {
+            request.addValue("Bearer \(token)", forHTTPHeaderField: "Authorization")
+        }
+        let (data, _) = try await URLSession.shared.data(for: request)
+        guard let json = try? JSONSerialization.jsonObject(with: data) as? [String: Any] else {
+            return nil
+        }
+        guard let link = json["link"] as? String, let fileName = json["file_name"] as?
+            String, let url = URL(string: link)
+        else {
+            return nil
+        }
+        return URLSubtitleInfo(subtitleID: String(fileID), name: fileName, url: url)
     }
 }
 
@@ -282,26 +397,17 @@ extension URL {
 
         let hash = offsets.map { offset -> String in
             file.seek(toFileOffset: offset)
-            return (file.readData(ofLength: 4096) as NSData).md5()
+            return file.readData(ofLength: 4096).md5()
         }.joined(separator: ";")
         return hash
     }
 }
 
-import CommonCrypto
-extension NSData {
+import CryptoKit
+
+public extension Data {
     func md5() -> String {
-        let digestLength = Int(CC_MD5_DIGEST_LENGTH)
-        let md5Buffer = UnsafeMutablePointer<CUnsignedChar>.allocate(capacity: digestLength)
-
-        CC_MD5(bytes, CC_LONG(length), md5Buffer)
-
-        let output = NSMutableString(capacity: Int(CC_MD5_DIGEST_LENGTH * 2))
-        for i in 0 ..< digestLength {
-            output.appendFormat("%02x", md5Buffer[i])
-        }
-
-        md5Buffer.deallocate()
-        return NSString(format: output) as String
+        let digestData = Insecure.MD5.hash(data: self)
+        return String(digestData.map { String(format: "%02hhx", $0) }.joined().prefix(32))
     }
 }
