@@ -5,18 +5,32 @@
 //  Created by kintan on 2022/1/29.
 //
 import AVFoundation
+import MediaPlayer
 import SwiftUI
 
-@available(iOS 16.0, macOS 13.0, tvOS 16.0, watchOS 9.0, *)
+@available(iOS 16.0, macOS 13.0, tvOS 16.0, *)
 public struct KSVideoPlayerView: View {
     private let subtitleDataSouce: SubtitleDataSouce?
-    private let onPlayerDisappear: ((KSPlayerLayer?) -> Void)?
-    @State private var delayItem: DispatchWorkItem?
-    @State private var overView = false
-    @StateObject private var playerCoordinator = KSVideoPlayer.Coordinator()
-    @Environment(\.dismiss) private var dismiss
+    @State
+    private var title: String
+    @StateObject
+    private var playerCoordinator: KSVideoPlayer.Coordinator
+    @Environment(\.dismiss)
+    private var dismiss
+    @FocusState
+    private var focusableField: FocusableField? {
+        willSet {
+            isDropdownShow = newValue == .info
+        }
+    }
+
     public let options: KSOptions
-    @State public var url: URL {
+    @State
+    private var isDropdownShow = false
+    @State
+    private var showVideoSetting = false
+    @State
+    public var url: URL {
         didSet {
             #if os(macOS)
             NSDocumentController.shared.noteNewRecentDocumentURL(url)
@@ -24,135 +38,121 @@ public struct KSVideoPlayerView: View {
         }
     }
 
-    @State var isMaskShow = true {
-        didSet {
-            if isMaskShow != oldValue {
-                if isMaskShow {
-                    delayItem?.cancel()
-                    // 播放的时候才自动隐藏
-                    guard playerCoordinator.state == .bufferFinished else { return }
-                    delayItem = DispatchWorkItem {
-                        isMaskShow = false
-                    }
-                    DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + KSOptions.animateDelayTimeInterval,
-                                                  execute: delayItem!)
-                }
-                #if os(macOS)
-                isMaskShow ? NSCursor.unhide() : NSCursor.setHiddenUntilMouseMoves(true)
-                if let window = playerCoordinator.playerLayer?.player.view?.window {
-                    window.standardWindowButton(.zoomButton)?.isHidden = !isMaskShow
-                    window.standardWindowButton(.closeButton)?.isHidden = !isMaskShow
-                    window.standardWindowButton(.miniaturizeButton)?.isHidden = !isMaskShow
-//                    window.standardWindowButton(.closeButton)?.superview?.isHidden = !isMaskShow
-                }
-                #endif
-            }
-        }
+    public init(url: URL, options: KSOptions, title: String? = nil, subtitleDataSouce: SubtitleDataSouce? = nil) {
+        self.init(coordinator: KSVideoPlayer.Coordinator(), url: url, options: options, title: title, subtitleDataSouce: subtitleDataSouce)
     }
 
-    public init(url: URL, options: KSOptions, subtitleDataSouce: SubtitleDataSouce? = nil, onPlayerDisappear: ((KSPlayerLayer?) -> Void)? = nil) {
+    public init(coordinator: KSVideoPlayer.Coordinator, url: URL, options: KSOptions, title: String? = nil, subtitleDataSouce: SubtitleDataSouce? = nil) {
         _url = .init(initialValue: url)
+        _playerCoordinator = .init(wrappedValue: coordinator)
+        _title = .init(initialValue: title ?? url.lastPathComponent)
         #if os(macOS)
         NSDocumentController.shared.noteNewRecentDocumentURL(url)
         #endif
         self.options = options
         self.subtitleDataSouce = subtitleDataSouce
-        self.onPlayerDisappear = onPlayerDisappear
     }
 
     public var body: some View {
         ZStack {
-            KSVideoPlayer(coordinator: playerCoordinator, url: url, options: options)
-                .onStateChanged { _, state in
-                    if state == .bufferFinished {
-                        isMaskShow = false
-                    } else {
-                        isMaskShow = true
-                    }
+            GeometryReader { proxy in
+                playView
+                HStack {
+                    Spacer()
+                    VideoSubtitleView(model: playerCoordinator.subtitleModel)
+                    Spacer()
                 }
-            #if canImport(UIKit)
-                .onSwipe { direction in
-                    isMaskShow = true
-                    if direction == .left {
-                        playerCoordinator.skip(interval: -15)
-                    } else if direction == .right {
-                        playerCoordinator.skip(interval: 15)
-                    }
-                }
-            #endif
-            #if !os(tvOS) && !os(macOS)
-                .onTapGesture {
-                isMaskShow.toggle()
-            }
-            #endif
-            .onAppear {
-                if let subtitleDataSouce {
-                    playerCoordinator.subtitleModel.addSubtitle(dataSouce: subtitleDataSouce)
-                }
-                #if os(macOS)
-                NSEvent.addLocalMonitorForEvents(matching: [.mouseMoved]) {
-                    isMaskShow = overView
-                    return $0
+                .padding()
+                controllerView(playerWidth: proxy.size.width)
+                #if os(tvOS)
+                if isDropdownShow {
+                    VideoSettingView(config: playerCoordinator, subtitleModel: playerCoordinator.subtitleModel, subtitleTitle: title)
+                        .focused($focusableField, equals: .info)
                 }
                 #endif
             }
-            .onDisappear {
-                delayItem?.cancel()
-                onPlayerDisappear?(playerCoordinator.playerLayer)
-                if let playerLayer = playerCoordinator.playerLayer {
-                    if !playerLayer.isPipActive {
-                        playerCoordinator.resetPlayer()
-                    }
-                }
-            }
-            .ignoresSafeArea()
-            VideoSubtitleView(model: playerCoordinator.subtitleModel)
-            VStack {
-                Spacer()
-                VStack {
-                    VideoControllerView(config: playerCoordinator)
-                    #if !os(iOS)
-                        .onMoveCommand { direction in
-                            isMaskShow = true
-                            #if os(macOS)
-                            switch direction {
-                            case .left:
-                                playerCoordinator.skip(interval: -15)
-                            case .right:
-                                playerCoordinator.skip(interval: 15)
-                            case .up:
-                                playerCoordinator.playerLayer?.player.playbackVolume += 1
-                            case .down:
-                                playerCoordinator.playerLayer?.player.playbackVolume -= 1
-                            @unknown default:
-                                break
-                            }
-                            #endif
-                        }
-                        .onExitCommand {
-                            if isMaskShow {
-                                isMaskShow = false
-                            } else {
-                                dismiss()
-                            }
-                        }
-                    #endif
-                    // 设置opacity为0，还是会去更新View。所以只能这样了
-                    if isMaskShow {
-                        VideoTimeShowView(config: playerCoordinator, model: playerCoordinator.timemodel)
-                    }
-                }
-                .padding()
-                .background(.black.opacity(0.2))
-                .clipShape(RoundedRectangle(cornerRadius: 5))
-                .opacity(isMaskShow ? 1 : 0)
-            }
         }
         .preferredColorScheme(.dark)
-        .background(Color.black)
-        .foregroundColor(.white)
+        .tint(.white)
         .persistentSystemOverlays(.hidden)
-        .toolbar(isMaskShow ? .visible : .hidden, for: .automatic)
+        .toolbar(.hidden, for: .automatic)
+        #if os(tvOS)
+            .onPlayPauseCommand {
+                if playerCoordinator.state.isPlaying {
+                    playerCoordinator.playerLayer?.pause()
+                } else {
+                    playerCoordinator.playerLayer?.play()
+                }
+            }
+            .onExitCommand {
+                if playerCoordinator.isMaskShow {
+                    playerCoordinator.isMaskShow = false
+                } else {
+                    switch focusableField {
+                    case .play:
+                        dismiss()
+                    default:
+                        focusableField = .play
+                    }
+                }
+            }
+        #endif
+    }
+
+    private var playView: some View {
+        KSVideoPlayer(coordinator: playerCoordinator, url: url, options: options)
+            .onStateChanged { playerLayer, state in
+                if state == .readyToPlay {
+                    if let movieTitle = playerLayer.player.dynamicInfo?.metadata["title"] {
+                        title = movieTitle
+                    }
+                }
+            }
+            .onBufferChanged { bufferedCount, consumeTime in
+                print("bufferedCount \(bufferedCount), consumeTime \(consumeTime)")
+            }
+        #if canImport(UIKit)
+            .onSwipe { _ in
+                playerCoordinator.isMaskShow = true
+            }
+        #endif
+            .ignoresSafeArea()
+            .onAppear {
+                focusableField = .play
+                if let subtitleDataSouce {
+                    playerCoordinator.subtitleModel.addSubtitle(dataSouce: subtitleDataSouce)
+                }
+                // 不要加这个，不然playerCoordinator无法释放，也可以在onDisappear调用removeMonitor释放
+                //                    #if os(macOS)
+                //                    NSEvent.addLocalMonitorForEvents(matching: [.mouseMoved]) {
+                //                        isMaskShow = overView
+                //                        return $0
+                //                    }
+                //                    #endif
+            }
+
+        #if os(iOS) || os(xrOS)
+            .navigationBarTitleDisplayMode(.inline)
+        #endif
+        #if !os(iOS)
+            .focusable(!playerCoordinator.isMaskShow)
+        .focused($focusableField, equals: .play)
+        #endif
+        #if !os(xrOS)
+            .onKeyPressLeftArrow {
+            playerCoordinator.skip(interval: -15)
+        }
+        .onKeyPressRightArrow {
+            playerCoordinator.skip(interval: 15)
+        }
+        .onKeyPressSapce {
+            if playerCoordinator.state.isPlaying {
+                playerCoordinator.playerLayer?.pause()
+            } else {
+                playerCoordinator.playerLayer?.play()
+            }
+        }
+        #endif
         #if os(macOS)
             .onTapGesture(count: 2) {
                 guard let view = playerCoordinator.playerLayer?.player.view else {
@@ -161,31 +161,127 @@ public struct KSVideoPlayerView: View {
                 view.window?.toggleFullScreen(nil)
                 view.needsLayout = true
                 view.layoutSubtreeIfNeeded()
+        }
+        .onExitCommand {
+            playerCoordinator.playerLayer?.player.view?.exitFullScreenMode()
+        }
+        .onMoveCommand { direction in
+            switch direction {
+            case .left:
+                playerCoordinator.skip(interval: -15)
+            case .right:
+                playerCoordinator.skip(interval: 15)
+            case .up:
+                playerCoordinator.playerLayer?.player.playbackVolume += 0.2
+            case .down:
+                playerCoordinator.playerLayer?.player.playbackVolume -= 0.2
+            @unknown default:
+                break
             }
-            .onExitCommand {
-                playerCoordinator.playerLayer?.player.view?.exitFullScreenMode()
+        }
+        #else
+        .onTapGesture {
+                playerCoordinator.isMaskShow.toggle()
             }
-        #elseif os(iOS)
-            .navigationBarTitleDisplayMode(.inline)
         #endif
-        #if !os(tvOS)
-        .onHover {
-            overView = $0
-            isMaskShow = overView
-        }
-        .onDrop(of: ["public.file-url"], isTargeted: nil) { providers -> Bool in
-            providers.first?.loadDataRepresentation(forTypeIdentifier: "public.file-url") { data, _ in
-                if let data, let path = NSString(data: data, encoding: 4), let url = URL(string: path as String) {
-                    openURL(url)
-                }
+        #if os(tvOS)
+            .onMoveCommand { direction in
+            switch direction {
+            case .left:
+                playerCoordinator.skip(interval: -15)
+            case .right:
+                playerCoordinator.skip(interval: 15)
+            case .up:
+                playerCoordinator.isMaskShow.toggle()
+            case .down:
+                focusableField = .info
+            @unknown default:
+                break
             }
-            return true
         }
+        #else
+        .onHover { _ in
+                playerCoordinator.isMaskShow = true
+            }
+            .onDrop(of: ["public.file-url"], isTargeted: nil) { providers -> Bool in
+                providers.first?.loadDataRepresentation(forTypeIdentifier: "public.file-url") { data, _ in
+                    if let data, let path = NSString(data: data, encoding: 4), let url = URL(string: path as String) {
+                        openURL(url)
+                    }
+                }
+                return true
+            }
         #endif
     }
 
+    private func controllerView(playerWidth: Double) -> some View {
+        VStack {
+            VideoControllerView(config: playerCoordinator, subtitleModel: playerCoordinator.subtitleModel, title: $title, volumeSliderSize: playerWidth / 4)
+            #if !os(xrOS)
+            // 设置opacity为0，还是会去更新View。所以只能这样了
+            if playerCoordinator.isMaskShow {
+                VideoTimeShowView(config: playerCoordinator, model: playerCoordinator.timemodel)
+                    .onAppear {
+                        focusableField = .controller
+                    }
+                    .onDisappear {
+                        focusableField = .play
+                    }
+            }
+            #endif
+        }
+        #if os(xrOS)
+        .ornament(visibility: playerCoordinator.isMaskShow ? .visible : .hidden, attachmentAnchor: .scene(.bottom)) {
+            ornamentView(playerWidth: playerWidth)
+        }
+        .sheet(isPresented: $showVideoSetting) {
+            NavigationStack {
+                VideoSettingView(config: playerCoordinator, subtitleModel: playerCoordinator.subtitleModel, subtitleTitle: title)
+            }
+            .buttonStyle(.plain)
+        }
+        #endif
+        .focused($focusableField, equals: .controller)
+        .opacity(playerCoordinator.isMaskShow ? 1 : 0)
+        .padding()
+    }
+
+    private func ornamentView(playerWidth: Double) -> some View {
+        VStack(alignment: .leading) {
+            KSVideoPlayerViewBuilder.titleView(title: title, config: playerCoordinator)
+            ornamentControlsView(playerWidth: playerWidth)
+        }
+        .frame(width: playerWidth / 1.5)
+        .buttonStyle(.plain)
+        .padding(.vertical, 24)
+        .padding(.horizontal, 36)
+        #if os(xrOS)
+            .glassBackgroundEffect()
+        #endif
+    }
+
+    private func ornamentControlsView(playerWidth _: Double) -> some View {
+        HStack {
+            KSVideoPlayerViewBuilder.playbackControlView(config: playerCoordinator, spacing: 16)
+            Spacer()
+            VideoTimeShowView(config: playerCoordinator, model: playerCoordinator.timemodel, timeFont: .title3.monospacedDigit())
+            Spacer()
+            Group {
+                KSVideoPlayerViewBuilder.contentModeButton(config: playerCoordinator)
+                KSVideoPlayerViewBuilder.subtitleButton(config: playerCoordinator)
+                KSVideoPlayerViewBuilder.playbackRateButton(playbackRate: $playerCoordinator.playbackRate)
+                KSVideoPlayerViewBuilder.infoButton(showVideoSetting: $showVideoSetting)
+            }
+            .font(.largeTitle)
+        }
+    }
+
+    fileprivate enum FocusableField {
+        case play, controller, info
+    }
+
     public func openURL(_ url: URL) {
-        runInMainqueue {
+        runOnMainThread {
             if url.isAudio || url.isMovie {
                 self.url = url
             } else {
@@ -196,112 +292,247 @@ public struct KSVideoPlayerView: View {
     }
 }
 
-@available(iOS 15, tvOS 15, macOS 12, *)
+extension View {
+    func onKeyPressLeftArrow(action: @escaping () -> Void) -> some View {
+        if #available(iOS 17.0, macOS 14.0, tvOS 17.0, *) {
+            return onKeyPress(.leftArrow) {
+                action()
+                return .handled
+            }
+        } else {
+            return self
+        }
+    }
+
+    func onKeyPressRightArrow(action: @escaping () -> Void) -> some View {
+        if #available(iOS 17.0, macOS 14.0, tvOS 17.0, *) {
+            return onKeyPress(.rightArrow) {
+                action()
+                return .handled
+            }
+        } else {
+            return self
+        }
+    }
+
+    func onKeyPressSapce(action: @escaping () -> Void) -> some View {
+        if #available(iOS 17.0, macOS 14.0, tvOS 17.0, *) {
+            return onKeyPress(.space) {
+                action()
+                return .handled
+            }
+        } else {
+            return self
+        }
+    }
+}
+
+@available(iOS 16, tvOS 16, macOS 13, *)
 struct VideoControllerView: View {
     @ObservedObject
     fileprivate var config: KSVideoPlayer.Coordinator
+    @ObservedObject
+    fileprivate var subtitleModel: SubtitleModel
+    @Binding
+    fileprivate var title: String
+    fileprivate var volumeSliderSize: Double?
     @State
     private var showVideoSetting = false
+    @Environment(\.dismiss)
+    private var dismiss
     public var body: some View {
-        HStack {
+        VStack {
+            #if os(tvOS)
+            Spacer()
             HStack {
-                Button {
-                    config.isMuted.toggle()
-                } label: {
-                    Image(systemName: config.isMuted ? "speaker.slash.fill" : "speaker.wave.2.fill")
+                Text(title)
+                    .lineLimit(2)
+                    .layoutPriority(3)
+                ProgressView()
+                    .opacity(config.state == .buffering ? 1 : 0)
+                Spacer()
+                    .layoutPriority(2)
+                HStack {
+                    Button {
+                        if config.state.isPlaying {
+                            config.playerLayer?.pause()
+                        } else {
+                            config.playerLayer?.play()
+                        }
+                    } label: {
+                        Image(systemName: config.state == .error ? "play.slash.fill" : (config.state.isPlaying ? "pause.circle.fill" : "play.circle.fill"))
+                    }
+                    if let audioTracks = config.playerLayer?.player.tracks(mediaType: .audio), !audioTracks.isEmpty {
+                        audioButton(audioTracks: audioTracks)
+                    }
+                    muteButton
+                    contentModeButton
+                    subtitleButton
+                    playbackRateButton
+                    //                pipButton
+                    infoButton
                 }
+                .font(.caption)
+            }
+            #else
+            HStack {
+                #if !os(xrOS)
                 Button {
-                    config.isScaleAspectFill.toggle()
+                    dismiss()
                 } label: {
-                    Image(systemName: config.isScaleAspectFill ? "rectangle.arrowtriangle.2.inward" : "rectangle.arrowtriangle.2.outward")
+                    Image(systemName: "x.circle.fill")
                 }
-                #if !os(tvOS) && !os(xrOS)
+                #if !os(tvOS)
                 if config.playerLayer?.player.allowsExternalPlayback == true {
                     AirPlayView().fixedSize()
                 }
                 #endif
-                ProgressView().opacity(config.state == .buffering ? 1 : 0)
+                #endif
+                Spacer()
+                if let audioTracks = config.playerLayer?.player.tracks(mediaType: .audio), !audioTracks.isEmpty {
+                    audioButton(audioTracks: audioTracks)
+                    #if os(xrOS)
+                        .aspectRatio(1, contentMode: .fit)
+                        .glassBackgroundEffect()
+                    #endif
+                }
+                muteButton
+                #if !os(xrOS)
+                contentModeButton
+                subtitleButton
+                #endif
             }
-            .font(.system(.title2))
+            Spacer()
+            #if !os(xrOS)
+            KSVideoPlayerViewBuilder.playbackControlView(config: config)
             Spacer()
             HStack {
-                Button {
-                    config.skip(interval: -15)
-                } label: {
-                    Image(systemName: "gobackward.15")
-                }
-                #if !os(tvOS)
-                .keyboardShortcut(.leftArrow, modifiers: .none)
-                #endif
-                Button {
-                    if config.state.isPlaying {
-                        config.playerLayer?.pause()
-                    } else {
-                        config.playerLayer?.play()
-                    }
-                } label: {
-                    Image(systemName: config.state == .error ? "play.slash.fill" : (config.state.isPlaying ? "pause.fill" : "play.fill"))
-                }
-                .padding(.horizontal)
-                #if !os(tvOS)
-                    .keyboardShortcut(.space, modifiers: .none)
-                #endif
-                Button {
-                    config.skip(interval: 15)
-                } label: {
-                    Image(systemName: "goforward.15")
-                }
-                #if !os(tvOS)
-                .keyboardShortcut(.rightArrow, modifiers: .none)
-                #endif
+                KSVideoPlayerViewBuilder.titleView(title: title, config: config)
+                Spacer()
+                playbackRateButton
+                pipButton
+                infoButton
             }
-            .font(.system(.largeTitle))
-            Spacer()
-            HStack {
-                Button {
-                    config.playerLayer?.isPipActive.toggle()
-                } label: {
-                    Image(systemName: config.playerLayer?.isPipActive ?? false ? "pip.exit" : "pip.enter")
-                }
-                Button {
-                    showVideoSetting.toggle()
-                } label: {
-                    Image(systemName: "ellipsis.circle")
-                }
-                #if !os(tvOS)
-                .keyboardShortcut("s", modifiers: [.command, .shift])
-                #endif
-            }
-            .font(.system(.title2))
+            #endif
+            #endif
         }
-        .sheet(isPresented: $showVideoSetting) {
-            VideoSettingView(config: config, subtitleModel: config.subtitleModel)
-        }
-        #if os(tvOS)
-//            .focusSection()
-        .onPlayPauseCommand {
-            if config.state.isPlaying {
-                config.playerLayer?.pause()
-            } else {
-                config.playerLayer?.play()
-            }
-        }
-        #else
+        #if !os(tvOS)
+        .font(.title)
         .buttonStyle(.borderless)
+        #endif
+        .sheet(isPresented: $showVideoSetting) {
+            VideoSettingView(config: config, subtitleModel: config.subtitleModel, subtitleTitle: title)
+        }
+    }
+
+    private var muteButton: some View {
+        #if os(xrOS)
+        HStack {
+            Slider(value: $config.playbackVolume, in: 0 ... 1)
+                .onChange(of: config.playbackVolume) { _, newValue in
+                    config.isMuted = newValue == 0
+                }
+                .frame(width: volumeSliderSize ?? 100)
+                .tint(.white.opacity(0.8))
+                .padding(.leading, 16)
+            KSVideoPlayerViewBuilder.muteButton(config: config)
+        }
+        .padding(16)
+        .glassBackgroundEffect()
+        #else
+        KSVideoPlayerViewBuilder.muteButton(config: config)
+        #endif
+    }
+
+    private var contentModeButton: some View {
+        KSVideoPlayerViewBuilder.contentModeButton(config: config)
+    }
+
+    private func audioButton(audioTracks: [MediaPlayerTrack]) -> some View {
+        MenuView(selection: Binding {
+            audioTracks.first { $0.isEnabled }?.trackID
+        } set: { value in
+            if let track = audioTracks.first(where: { $0.trackID == value }) {
+                config.playerLayer?.player.select(track: track)
+            }
+        }) {
+            ForEach(audioTracks, id: \.trackID) { track in
+                Text(track.description).tag(track.trackID as Int32?)
+            }
+        } label: {
+            Image(systemName: "waveform.circle.fill")
+            #if os(xrOS)
+                .padding()
+                .clipShape(Circle())
+            #endif
+        }
+    }
+
+    private var subtitleButton: some View {
+        KSVideoPlayerViewBuilder.subtitleButton(config: config)
+    }
+
+    private var playbackRateButton: some View {
+        KSVideoPlayerViewBuilder.playbackRateButton(playbackRate: $config.playbackRate)
+    }
+
+    private var pipButton: some View {
+        Button {
+            config.playerLayer?.isPipActive.toggle()
+        } label: {
+            Image(systemName: "rectangle.on.rectangle.circle.fill")
+        }
+    }
+
+    private var infoButton: some View {
+        KSVideoPlayerViewBuilder.infoButton(showVideoSetting: $showVideoSetting)
+    }
+}
+
+@available(iOS 15, tvOS 16, macOS 12, *)
+public struct MenuView<Label, SelectionValue, Content>: View where Label: View, SelectionValue: Hashable, Content: View {
+    public let selection: Binding<SelectionValue>
+    @ViewBuilder
+    public let content: () -> Content
+    @ViewBuilder
+    public let label: () -> Label
+    @State
+    private var showMenu = false
+    public var body: some View {
+        #if os(tvOS)
+        Picker(selection: selection, content: content, label: label)
+            .pickerStyle(.navigationLink)
+            .frame(height: 50)
+        #else
+        Menu {
+            Picker(selection: selection) {
+                content()
+            } label: {
+                EmptyView()
+            }
+            .pickerStyle(.inline)
+        } label: {
+            // menu 里面的label无法调整大小
+            label()
+        }
+        .menuIndicator(.hidden)
         #endif
     }
 }
 
 @available(iOS 15, tvOS 15, macOS 12, *)
 struct VideoTimeShowView: View {
-    @ObservedObject fileprivate var config: KSVideoPlayer.Coordinator
-    @ObservedObject fileprivate var model: ControllerTimeModel
+    @ObservedObject
+    fileprivate var config: KSVideoPlayer.Coordinator
+    @ObservedObject
+    fileprivate var model: ControllerTimeModel
+    fileprivate var timeFont: Font?
     public var body: some View {
         if config.timemodel.totalTime == 0 {
             Text("Live Streaming")
         } else {
             HStack {
-                Text(model.currentTime.toString(for: .minOrHour)).font(.caption2.monospacedDigit())
+                Text(model.currentTime.toString(for: .minOrHour)).font(timeFont ?? .caption2.monospacedDigit())
                 Slider(value: Binding {
                     Double(model.currentTime)
                 } set: { newValue, _ in
@@ -314,8 +545,12 @@ struct VideoTimeShowView: View {
                     }
                 }
                 .frame(maxHeight: 20)
-                Text((model.totalTime).toString(for: .minOrHour)).font(.caption2.monospacedDigit())
+                #if os(xrOS)
+                    .tint(.white.opacity(0.8))
+                #endif
+                Text((model.totalTime).toString(for: .minOrHour)).font(timeFont ?? .caption2.monospacedDigit())
             }
+            .font(.system(.title2))
         }
     }
 }
@@ -326,123 +561,164 @@ extension EventModifiers {
 
 @available(iOS 16, tvOS 16, macOS 13, *)
 struct VideoSubtitleView: View {
-    @ObservedObject fileprivate var model: SubtitleModel
+    @ObservedObject
+    fileprivate var model: SubtitleModel
     var body: some View {
+        ZStack {
+            ForEach(model.parts) { part in
+                part.subtitleView
+            }
+        }
+    }
+
+    fileprivate static func imageView(_ image: UIImage) -> some View {
+        #if enableFeatureLiveText && canImport(VisionKit) && !targetEnvironment(simulator)
+        if #available(macCatalyst 17.0, *) {
+            return LiveTextImage(uiImage: image)
+        } else {
+            return Image(uiImage: image)
+                .resizable()
+        }
+        #else
+        return Image(uiImage: image)
+            .resizable()
+        #endif
+    }
+}
+
+private extension SubtitlePart {
+    @available(iOS 16, tvOS 16, macOS 13, *)
+    @MainActor
+    var subtitleView: some View {
         VStack {
-            if let image = model.part?.image {
+            if let image {
                 Spacer()
                 GeometryReader { geometry in
                     let fitRect = image.fitRect(geometry.size)
-                    imageView(image)
+                    VideoSubtitleView.imageView(image)
                         .offset(CGSize(width: fitRect.origin.x, height: fitRect.origin.y))
                         .frame(width: fitRect.size.width, height: fitRect.size.height)
                 }
                 // 不能加scaledToFit。不然的话图片的缩放比率会有问题。
 //                .scaledToFit()
                 .padding()
-            }
-            if let text = model.part?.text {
-                if SubtitleModel.textYAlign == .bottom {
+            } else if let text {
+                let textPosition = textPosition ?? SubtitleModel.textPosition
+                if textPosition.verticalAlign == .bottom || textPosition.verticalAlign == .center {
                     Spacer()
                 }
                 Text(AttributedString(text))
                     .font(Font(SubtitleModel.textFont))
                     .shadow(color: .black.opacity(0.9), radius: 1, x: 1, y: 1)
                     .foregroundColor(SubtitleModel.textColor)
-                    .background(SubtitleModel.textBackgroundColor)
-                    .multilineTextAlignment(SubtitleModel.textXAlign)
-                    .padding(SubtitleModel.edgeInsets)
                     .italic(SubtitleModel.textItalic)
+                    .background(SubtitleModel.textBackgroundColor)
+                    .multilineTextAlignment(.center)
+                    .alignmentGuide(textPosition.horizontalAlign) {
+                        $0[.leading]
+                    }
+                    .padding(textPosition.edgeInsets)
                 #if !os(tvOS)
                     .textSelection(.enabled)
                 #endif
-                if SubtitleModel.textYAlign == .top {
+                if textPosition.verticalAlign == .top || textPosition.verticalAlign == .center {
                     Spacer()
                 }
             }
         }
     }
-
-    private func imageView(_ image: UIImage) -> some View {
-        #if os(tvOS)
-        return Image(uiImage: image)
-            .resizable()
-        #else
-        return LiveTextImage(uiImage: image)
-        #endif
-    }
 }
 
-@available(iOS 15, tvOS 15, macOS 12, *)
+@available(iOS 16, tvOS 16, macOS 13, *)
 struct VideoSettingView: View {
     @ObservedObject
     fileprivate var config: KSVideoPlayer.Coordinator
     @ObservedObject
     fileprivate var subtitleModel: SubtitleModel
+    @State
+    fileprivate var subtitleTitle: String
     @Environment(\.dismiss)
     private var dismiss
+
     var body: some View {
-        Form {
-            Picker(selection: $config.playbackRate) {
-                ForEach([0.5, 1.0, 1.25, 1.5, 2.0] as [Float]) { value in
-                    // 需要有一个变量text。不然会自动帮忙加很多0
-                    let text = "\(value) x"
-                    Text(text).tag(value)
-                }
-            } label: {
-                Label("Playback Speed", systemImage: "speedometer")
-            }
-            if config.audioTracks.count > 0 {
+        PlatformView {
+            let videoTracks = config.playerLayer?.player.tracks(mediaType: .video)
+            if let videoTracks, !videoTracks.isEmpty {
                 Picker(selection: Binding {
-                    config.selectedAudioTrack?.trackID
+                    videoTracks.first { $0.isEnabled }?.trackID
                 } set: { value in
-                    config.selectedAudioTrack = config.audioTracks.first { $0.trackID == value }
+                    if let track = videoTracks.first(where: { $0.trackID == value }) {
+                        config.playerLayer?.player.select(track: track)
+                    }
                 }) {
-                    ForEach(config.audioTracks, id: \.trackID) { track in
+                    ForEach(videoTracks, id: \.trackID) { track in
                         Text(track.description).tag(track.trackID as Int32?)
                     }
                 } label: {
-                    Label("Audio track", systemImage: "waveform")
+                    Label("Video Track", systemImage: "video.fill")
                 }
+                LabeledContent("Video Type", value: (videoTracks.first { $0.isEnabled }?.dynamicRange ?? .sdr).description)
             }
-            if config.videoTracks.count > 0 {
-                Picker(selection: Binding {
-                    config.selectedVideoTrack?.trackID
-                } set: { value in
-                    config.selectedVideoTrack = config.videoTracks.first { $0.trackID == value }
-                }) {
-                    ForEach(config.videoTracks, id: \.trackID) { track in
-                        Text(track.description).tag(track.trackID as Int32?)
-                    }
-                } label: {
-                    Label("Video track", systemImage: "video.fill")
-                }
+            TextField("Sutitle delay", value: $subtitleModel.subtitleDelay, format: .number)
+            TextField("Title", text: $subtitleTitle)
+            Button("Search Sutitle") {
+                subtitleModel.searchSubtitle(query: subtitleTitle, languages: ["zh-cn"])
             }
-            if config.subtitleModel.subtitleInfos.count > 0 {
-                Picker(selection: Binding {
-                    subtitleModel.selectedSubtitleInfo?.subtitleID
-                } set: { value in
-                    subtitleModel.selectedSubtitleInfo = subtitleModel.subtitleInfos.first { $0.subtitleID == value }
-                }) {
-                    Text("Off").tag(nil as String?)
-                    ForEach(subtitleModel.subtitleInfos, id: \.subtitleID) { track in
-                        Text(track.name).tag(track.subtitleID as String?)
-                    }
-                } label: {
-                    Label("Sutitle", systemImage: "captions.bubble")
-                }
-                TextField("Sutitle delay", value: $subtitleModel.subtitleDelay, format: .number)
+            LabeledContent("Stream Type", value: (videoTracks?.first { $0.isEnabled }?.fieldOrder ?? .progressive).description)
+            if let dynamicInfo = config.playerLayer?.player.dynamicInfo {
+                DynamicInfoView(dynamicInfo: dynamicInfo)
+            }
+            if let fileSize = config.playerLayer?.player.fileSize, fileSize > 0 {
+                LabeledContent("File Size", value: fileSize.kmFormatted + "B")
             }
         }
-        .padding()
-        #if os(macOS)
-            .toolbar {
-                Button("Done") {
-                    dismiss()
-                }
-                .keyboardShortcut(.defaultAction)
+        #if os(macOS) || targetEnvironment(macCatalyst) || os(xrOS)
+        .toolbar {
+            Button("Done") {
+                dismiss()
             }
+            .keyboardShortcut(.defaultAction)
+        }
         #endif
+    }
+}
+
+@available(iOS 16, tvOS 16, macOS 13, *)
+public struct DynamicInfoView: View {
+    @ObservedObject
+    fileprivate var dynamicInfo: DynamicInfo
+    public var body: some View {
+        LabeledContent("Display FPS", value: dynamicInfo.displayFPS, format: .number)
+        LabeledContent("Audio Video sync", value: dynamicInfo.audioVideoSyncDiff, format: .number)
+        LabeledContent("Dropped Frames", value: dynamicInfo.droppedVideoFrameCount + dynamicInfo.droppedVideoPacketCount, format: .number)
+        LabeledContent("Bytes Read", value: dynamicInfo.bytesRead.kmFormatted + "B")
+        LabeledContent("Audio bitrate", value: dynamicInfo.audioBitrate.kmFormatted + "bps")
+        LabeledContent("Video bitrate", value: dynamicInfo.videoBitrate.kmFormatted + "bps")
+    }
+}
+
+@available(iOS 15, tvOS 16, macOS 12, *)
+public struct PlatformView<Content: View>: View {
+    private let content: () -> Content
+    public var body: some View {
+        #if os(tvOS)
+        ScrollView {
+            content()
+                .padding()
+        }
+        .pickerStyle(.navigationLink)
+        #else
+        Form {
+            content()
+        }
+        #if os(macOS)
+        .padding()
+        #endif
+        #endif
+    }
+
+    public init(@ViewBuilder content: @escaping () -> Content) {
+        self.content = content
     }
 }
 
@@ -450,7 +726,7 @@ struct VideoSettingView: View {
 struct KSVideoPlayerView_Previews: PreviewProvider {
     static var previews: some View {
         let url = URL(string: "http://clips.vorwaerts-gmbh.de/big_buck_bunny.mp4")!
-        KSVideoPlayerView(url: url, options: KSOptions())
+        KSVideoPlayerView(coordinator: KSVideoPlayer.Coordinator(), url: url, options: KSOptions())
     }
 }
 
