@@ -10,7 +10,8 @@ import SwiftUI
 
 struct FilesView: View {
     @FetchRequest(
-        sortDescriptors: [NSSortDescriptor(keyPath: \M3UModel.name, ascending: true)]
+        sortDescriptors: [NSSortDescriptor(keyPath: \M3UModel.name, ascending: true)],
+        predicate: NSPredicate(format: "m3uURL != nil && name != nil")
     )
     private var m3uModels: FetchedResults<M3UModel>
     @EnvironmentObject
@@ -18,15 +19,22 @@ struct FilesView: View {
     @State
     private var addM3U = false
     @State
+    private var openFileImport = false
+    @State
     private var nameFilter: String = ""
     var body: some View {
         let models = m3uModels.filter { model in
             var isIncluded = true
-            if nameFilter.count > 0 {
+            if !nameFilter.isEmpty {
                 isIncluded = model.name!.contains(nameFilter)
             }
             return isIncluded
         }
+        #if os(tvOS)
+        HStack {
+            toolbarView
+        }
+        #endif
         List(models, id: \.self, selection: $appModel.activeM3UModel) { model in
             #if os(tvOS)
             NavigationLink(value: model) {
@@ -37,26 +45,42 @@ struct FilesView: View {
             #endif
         }
         .searchable(text: $nameFilter)
-        .toolbar {
-            Button {
-                addM3U = true
-            } label: {
-                Label("Add M3U", systemImage: "plus.app.fill")
-            }
-        }
         .sheet(isPresented: $addM3U) {
             AddM3UView()
         }
+        #if !os(tvOS)
+        .toolbar {
+            toolbarView
+        }
+        .fileImporter(isPresented: $openFileImport, allowedContentTypes: [.data]) { result in
+            guard let url = try? result.get() else {
+                return
+            }
+            appModel.addM3U(url: url)
+        }
+        #endif
     }
 
-    private func cellView(model: M3UModel) -> some View {
-        #if os(tvOS)
-        NavigationLink(value: model) {
-            M3UView(model: model)
+    private var toolbarView: some View {
+        Group {
+            Button {
+                openFileImport = true
+            } label: {
+                Label("Add Local M3U", systemImage: "plus.rectangle.on.folder.fill")
+            }
+            #if !os(tvOS)
+            .keyboardShortcut("o")
+            #endif
+            Button {
+                addM3U = true
+            } label: {
+                Label("Add Remote M3U", systemImage: "plus.app.fill")
+            }
+            #if !os(tvOS)
+            .keyboardShortcut("o", modifiers: [.command, .shift])
+            #endif
         }
-        #else
-        M3UView(model: model)
-        #endif
+        .labelStyle(.titleAndIcon)
     }
 }
 
@@ -65,25 +89,27 @@ struct M3UView: View {
     var model: M3UModel
     var body: some View {
         VStack(alignment: .leading) {
-            Text(model.name!)
+            Text(model.name ?? "")
                 .font(.title2)
                 .foregroundColor(.primary)
             Text("total \(model.count) channels")
                 .font(.callout)
                 .foregroundColor(.secondary)
-            Text(model.m3uURL!.description)
+            Text(model.m3uURL?.description ?? "")
                 .font(.callout)
                 .foregroundColor(.secondary)
         }
         .contextMenu {
-            Button {
-                model.managedObjectContext?.delete(model)
-            } label: {
-                Label("Delete", systemImage: "trash.fill")
+            if PersistenceController.shared.container.canUpdateRecord(forManagedObjectWith: model.objectID) {
+                Button {
+                    model.delete()
+                } label: {
+                    Label("Delete", systemImage: "trash.fill")
+                }
             }
             Button {
                 Task {
-                    await _ = model.parsePlaylist(refresh: true)
+                    try? await _ = model.parsePlaylist()
                 }
             } label: {
                 Label("Refresh", systemImage: "arrow.clockwise.circle")
@@ -105,8 +131,15 @@ struct M3UView: View {
 }
 
 struct AddM3UView: View {
-    @State private var url = ""
-    @State private var name = ""
+    #if DEBUG && targetEnvironment(simulator)
+    @State
+    private var url = "https://raw.githubusercontent.com/kingslay/TestVideo/main/TestVideo.m3u"
+    #else
+    @State
+    private var url = ""
+    #endif
+    @State
+    private var name = ""
     @EnvironmentObject private var appModel: APPModel
     @Environment(\.dismiss) private var dismiss
     var body: some View {
@@ -120,14 +153,14 @@ struct AddM3UView: View {
                 Button("Done") {
                     if let url = URL(string: url.trimmingCharacters(in: NSMutableCharacterSet.whitespacesAndNewlines)) {
                         let name = name.trimmingCharacters(in: NSMutableCharacterSet.whitespacesAndNewlines)
-                        appModel.addM3U(url: url, name: name.count == 0 ? nil : name)
+                        appModel.addM3U(url: url, name: name.isEmpty ? nil : name)
                     }
                     dismiss()
                 }
                 #if !os(tvOS)
                 .keyboardShortcut(.defaultAction)
                 #endif
-                #if os(macOS)
+                #if os(macOS) || targetEnvironment(macCatalyst)
                 Button("Cancel") {
                     dismiss()
                 }
