@@ -29,6 +29,8 @@ struct SettingView: View {
                 }
             }
         }
+        .preferredColorScheme(.dark)
+        .background(Color.black)
     }
 }
 
@@ -43,12 +45,16 @@ struct SettingGeneralView: View {
 }
 
 struct SettingAudioView: View {
-    @Default(\.isUseAudioRenderer)
-    private var isUseAudioRenderer
+    @Default(\.audioPlayerType)
+    private var audioPlayerType
     init() {}
     var body: some View {
         Form {
-            Toggle("Use Audio Renderer", isOn: $isUseAudioRenderer)
+            Picker("audio Player Type", selection: $audioPlayerType) {
+                Text("AUGraph").tag(NSStringFromClass(AudioGraphPlayer.self))
+                Text("AudioUnit").tag(NSStringFromClass(AudioUnitPlayer.self))
+                Text("AVAudioEngine").tag(NSStringFromClass(AudioEnginePlayer.self))
+            }
         }
     }
 }
@@ -56,13 +62,26 @@ struct SettingAudioView: View {
 struct SettingVideoView: View {
     @Default(\.hardwareDecode)
     private var hardwareDecode
+    @Default(\.asynchronousDecompression)
+    private var asynchronousDecompression
     @Default(\.isUseDisplayLayer)
     private var isUseDisplayLayer
-
+    @Default(\.displayCriteriaFormatDescriptionEnabled)
+    private var displayCriteriaFormatDescriptionEnabled
+    @Default(\.yadifMode)
+    private var yadifMode
     var body: some View {
         Form {
             Toggle("Hardware decoder", isOn: $hardwareDecode)
+            Toggle("Asynchronous Decompression", isOn: $asynchronousDecompression)
             Toggle("Use DisplayLayer", isOn: $isUseDisplayLayer)
+            Toggle("Enable FormatDescription DisplayCriteria ", isOn: $displayCriteriaFormatDescriptionEnabled)
+            Picker("yadif Mode", selection: $yadifMode) {
+                Text("yadif").tag(0)
+                Text("yadif_2x").tag(1)
+                Text("yadif_spatial_skip").tag(2)
+                Text("yadif_2x_spatial_skip").tag(3)
+            }
         }
     }
 }
@@ -70,7 +89,7 @@ struct SettingVideoView: View {
 struct SettingSubtitleView: View {
     @Default(\.textFontSize)
     private var textFontSize
-    @Default(\.textItalic)
+    @Default(\.textBold)
     private var textBold
     @Default(\.textItalic)
     private var textItalic
@@ -78,15 +97,16 @@ struct SettingSubtitleView: View {
     private var textColor
     @Default(\.textBackgroundColor)
     private var textBackgroundColor
-    @Default(\.textXAlign)
-    private var textXAlign
-    @Default(\.textYAlign)
-    private var textYAlign
-    @Default(\.textXMargin)
-    private var textXMargin
-    @Default(\.textYMargin)
-    private var textYMargin
-
+    @Default(\.verticalAlign)
+    private var verticalAlign
+    @Default(\.horizontalAlign)
+    private var horizontalAlign
+    @Default(\.leftMargin)
+    private var leftMargin
+    @Default(\.rightMargin)
+    private var rightMargin
+    @Default(\.verticalMargin)
+    private var verticalMargin
     var body: some View {
         Form {
             Section("Position") {
@@ -104,27 +124,33 @@ struct SettingSubtitleView: View {
                 #endif
             }
             Section("Position") {
-                Picker("Align X:", selection: $textXAlign) {
-                    ForEach([TextAlignment.leading, .center, .trailing]) { value in
+                Picker("Align X:", selection: $horizontalAlign) {
+                    ForEach([HorizontalAlignment.leading, .center, .trailing]) { value in
                         Text(value.rawValue).tag(value)
                     }
                 }
-                Picker("Align Y:", selection: $textYAlign) {
+                Picker("Align Y:", selection: $verticalAlign) {
                     ForEach([VerticalAlignment.top, .center, .bottom]) { value in
                         Text(value.rawValue).tag(value)
                     }
                 }
                 HStack {
                     #if os(iOS)
-                    Text("Margin X:")
+                    Text("Margin Left:")
                     #endif
-                    TextField("Margin X:", value: $textXMargin, format: .number)
+                    TextField("Margin Left:", value: $leftMargin, format: .number)
                 }
                 HStack {
                     #if os(iOS)
-                    Text("Margin Y:")
+                    Text("Margin Right:")
                     #endif
-                    TextField("Margin Y:", value: $textYMargin, format: .number)
+                    TextField("Margin Right:", value: $rightMargin, format: .number)
+                }
+                HStack {
+                    #if os(iOS)
+                    Text("Margin Vertical:")
+                    #endif
+                    TextField("Margin Vertical:", value: $verticalMargin, format: .number)
                 }
             }
         }
