@@ -1,5 +1,5 @@
 //
-//  ContentView.swift
+//  URLImportView.swift
 //  Demo
 //
 //  Created by kintan on 2020/3/22.
@@ -22,7 +22,7 @@ struct URLImportView: View {
             Section {
                 TextField("URL:", text: $playURL)
                 Toggle("Remember URL", isOn: $rememberURL)
-                if historyURLs.count > 0 {
+                if !historyURLs.isEmpty {
                     Picker("History URL", selection: $playURL) {
                         Text("None").tag("")
                         ForEach(historyURLs) {
@@ -43,11 +43,11 @@ struct URLImportView: View {
                 Button("Done") {
                     dismiss()
                     let urlString = playURL.trimmingCharacters(in: NSMutableCharacterSet.whitespacesAndNewlines)
-                    if urlString.count > 0, var components = URLComponents(string: urlString) {
-                        if username.count > 0 {
+                    if !urlString.isEmpty, var components = URLComponents(string: urlString) {
+                        if !username.isEmpty {
                             components.user = username
                         }
-                        if password.count > 0 {
+                        if !password.isEmpty {
                             components.password = password
                         }
                         if let url = components.url {
@@ -68,7 +68,7 @@ struct URLImportView: View {
                 #if !os(tvOS)
                 .keyboardShortcut(.defaultAction)
                 #endif
-                #if os(macOS)
+                #if os(macOS) || targetEnvironment(macCatalyst)
                 Button("Cancel") {
                     dismiss()
                 }
