@@ -6,12 +6,18 @@
 //
 
 import AVFoundation
+#if os(tvOS) || os(xrOS)
+import DisplayCriteria
+#endif
 import OSLog
 
+#if canImport(UIKit)
+import UIKit
+#endif
 open class KSOptions {
-    //    public static let shared = KSOptions()
     /// 最低缓存视频时间
-    @Published public var preferredForwardBufferDuration = KSOptions.preferredForwardBufferDuration
+    @Published
+    public var preferredForwardBufferDuration = KSOptions.preferredForwardBufferDuration
     /// 最大缓存视频时间
     public var maxBufferDuration = KSOptions.maxBufferDuration
     /// 是否开启秒开
@@ -30,21 +36,22 @@ open class KSOptions {
      AVSEEK_FLAG_ANY: 4
      AVSEEK_FLAG_FRAME: 8
      */
-    public var seekFlags = Int32(0)
+    public var seekFlags = Int32(1)
     // ffmpeg only cache http
     public var cache = false
+    //  record stream
     public var outputURL: URL?
-    public var display = DisplayEnum.plane
     public var avOptions = [String: Any]()
     public var formatContextOptions = [String: Any]()
     public var decoderOptions = [String: Any]()
     public var probesize: Int64?
     public var maxAnalyzeDuration: Int64?
     public var lowres = UInt8(0)
+    public var nobuffer = false
+    public var codecLowDelay = false
     public var startPlayTime: TimeInterval = 0
     public var startPlayRate: Float = 1.0
-    public var registerRemoteControll: Bool = true//默认支持来自系统控制中心的控制
-
+    public var registerRemoteControll: Bool = true // 默认支持来自系统控制中心的控制
     public var referer: String? {
         didSet {
             if let referer {
@@ -55,25 +62,23 @@ open class KSOptions {
         }
     }
 
-    public var userAgent: String? {
+    public var userAgent: String? = "KSPlayer" {
         didSet {
-            if let userAgent {
-                formatContextOptions["user_agent"] = userAgent
-            } else {
-                formatContextOptions["user_agent"] = nil
-            }
+            formatContextOptions["user_agent"] = userAgent
         }
     }
 
     // audio
-    public var audioDelay = 0.0 // s
     public var audioFilters = [String]()
     public var syncDecodeAudio = false
+    // Locale(identifier: "en-US") Locale(identifier: "zh-CN")
+    public var audioLocale: Locale?
     // sutile
     public var autoSelectEmbedSubtitle = true
-    public var subtitleDisable = false
     public var isSeekImageSubtitle = false
     // video
+    public var display = DisplayEnum.plane
+    public var videoDelay = 0.0 // s
     public var autoDeInterlace = false
     public var autoRotate = true
     public var destinationDynamicRange: DynamicRange?
@@ -81,9 +86,12 @@ open class KSOptions {
     public var videoFilters = [String]()
     public var syncDecodeVideo = false
     public var hardwareDecode = KSOptions.hardwareDecode
-    public var asynchronousDecompression = true
+    public var asynchronousDecompression = KSOptions.asynchronousDecompression
     public var videoDisable = false
-    public var canStartPictureInPictureAutomaticallyFromInline = true
+    public var canStartPictureInPictureAutomaticallyFromInline = KSOptions.canStartPictureInPictureAutomaticallyFromInline
+    public var automaticWindowResize = true
+    @Published
+    public var videoInterlacingType: VideoInterlacingType?
     private var videoClockDelayCount = 0
 
     public internal(set) var formatName = ""
@@ -98,21 +106,30 @@ open class KSOptions {
     public internal(set) var readVideoTime = 0.0
     public internal(set) var decodeAudioTime = 0.0
     public internal(set) var decodeVideoTime = 0.0
-    var audioFormat = AVAudioFormat(standardFormatWithSampleRate: 44100, channelLayout: AVAudioChannelLayout(layoutTag: kAudioChannelLayoutTag_Stereo)!)
     public init() {
+        formatContextOptions["user_agent"] = userAgent
         // 参数的配置可以参考protocols.texi 和 http.c
-        formatContextOptions["auto_convert"] = 0
-        formatContextOptions["fps_probe_size"] = 3
+        // 这个一定要，不然有的流就会判断不准FieldOrder
+        formatContextOptions["scan_all_pmts"] = 1
+        // ts直播流需要加这个才能一直直播下去，不然播放一小段就会结束了。
         formatContextOptions["reconnect"] = 1
-        // 开启这个，纯ipv6地址会无法播放。
-//        formatContextOptions["reconnect_at_eof"] = 1
         formatContextOptions["reconnect_streamed"] = 1
-        formatContextOptions["reconnect_on_network_error"] = 1
-
+        // 这个是用来开启http的链接复用（keep-alive）。vlc默认是打开的，所以这边也默认打开。
+        formatContextOptions["multiple_requests"] = 1
+        // 下面是用来处理秒开的参数，有需要的自己打开。默认不开，不然在播放某些特殊的ts直播流会频繁卡顿。
+//        formatContextOptions["auto_convert"] = 0
+//        formatContextOptions["fps_probe_size"] = 3
+//        formatContextOptions["rw_timeout"] = 10_000_000
+//        formatContextOptions["max_analyze_duration"] = 300 * 1000
+        // 默认情况下允许所有协议，只有嵌套协议才需要指定这个协议子集，例如m3u8里面有http。
+//        formatContextOptions["protocol_whitelist"] = "file,http,https,tcp,tls,crypto,async,cache,data,httpproxy"
+        // 开启这个，纯ipv6地址会无法播放。并且有些视频结束了，但还会一直尝试重连。所以这个值默认不设置
+//        formatContextOptions["reconnect_at_eof"] = 1
+        // 开启这个，会导致tcp Failed to resolve hostname 还会一直重试
+//        formatContextOptions["reconnect_on_network_error"] = 1
         // There is total different meaning for 'listen_timeout' option in rtmp
         // set 'listen_timeout' = -1 for rtmp、rtsp
 //        formatContextOptions["listen_timeout"] = 3
-        formatContextOptions["rw_timeout"] = 10_000_000
         decoderOptions["threads"] = "auto"
         decoderOptions["refcounted_frames"] = "1"
     }
@@ -146,20 +163,16 @@ open class KSOptions {
 
     // 缓冲算法函数
     open func playable(capacitys: [CapacityProtocol], isFirst: Bool, isSeek: Bool) -> LoadingState {
-        let packetCount = capacitys.map(\.packetCount).max() ?? 0
-        let frameCount = capacitys.map(\.frameCount).max() ?? 0
+        let packetCount = capacitys.map(\.packetCount).min() ?? 0
+        let frameCount = capacitys.map(\.frameCount).min() ?? 0
         let isEndOfFile = capacitys.allSatisfy(\.isEndOfFile)
-        let loadedTime = capacitys.map { TimeInterval($0.packetCount + $0.frameCount) / TimeInterval($0.fps) }.max() ?? 0
-        let progress = loadedTime * 100.0 / preferredForwardBufferDuration
+        let loadedTime = capacitys.map(\.loadedTime).min() ?? 0
+        let progress = preferredForwardBufferDuration == 0 ? 100 : loadedTime * 100.0 / preferredForwardBufferDuration
         let isPlayable = capacitys.allSatisfy { capacity in
             if capacity.isEndOfFile && capacity.packetCount == 0 {
                 return true
             }
-            // 处理视频轨道一致没有值的问题(纯音频)
-            if capacity.mediaType == .video && capacity.frameCount == 0 && capacity.packetCount == 0 {
-                return true
-            }
-            guard capacity.frameCount >= capacity.frameMaxCount >> 2 else {
+            guard capacity.frameCount >= 2 else {
                 return false
             }
             if capacity.isEndOfFile {
@@ -173,12 +186,12 @@ open class KSOptions {
                 if capacity.mediaType == .audio || isSecondOpen {
                     if isFirst {
                         return true
-                    } else if isSeek, capacity.packetCount >= Int(capacity.fps) {
-                        return true
+                    } else {
+                        return capacity.loadedTime >= self.preferredForwardBufferDuration / 2
                     }
                 }
             }
-            return capacity.packetCount + capacity.frameCount >= Int(capacity.fps * Float(preferredForwardBufferDuration))
+            return capacity.loadedTime >= self.preferredForwardBufferDuration
         }
         return LoadingState(loadedTime: loadedTime, progress: progress, packetCount: packetCount,
                             frameCount: frameCount, isEndOfFile: isEndOfFile, isPlayable: isPlayable,
@@ -206,25 +219,30 @@ open class KSOptions {
     }
 
     ///  wanted video stream index, or nil for automatic selection
-    /// - Parameter : video bitRate
-    /// - Returns: The index of the bitRates
-    open func wantedVideo(bitRates _: [Int64]) -> Int? {
+    /// - Parameter : video track
+    /// - Returns: The index of the track
+    open func wantedVideo(tracks _: [MediaPlayerTrack]) -> Int? {
         nil
     }
 
     /// wanted audio stream index, or nil for automatic selection
-    /// - Parameter :  audio bitRate and language
-    /// - Returns: The index of the infos
-    open func wantedAudio(infos _: [(bitRate: Int64, language: String?)]) -> Int? {
+    /// - Parameter :  audio track
+    /// - Returns: The index of the track
+    open func wantedAudio(tracks _: [MediaPlayerTrack]) -> Int? {
         nil
     }
 
-    open func videoFrameMaxCount(fps: Float) -> Int {
-        Int(ceil(fps)) >> 1
+    open func videoFrameMaxCount(fps _: Float, naturalSize _: CGSize, isLive: Bool) -> UInt8 {
+        isLive ? 4 : 16
     }
 
-    open func audioFrameMaxCount(fps _: Float, channels: Int) -> Int {
-        (16 * max(channels, 1)) >> 1
+    open func audioFrameMaxCount(fps: Float, channelCount: Int) -> UInt8 {
+        let count = (Int(fps) * channelCount) >> 2
+        if count >= UInt8.max {
+            return UInt8.max
+        } else {
+            return UInt8(count)
+        }
     }
 
     /// customize dar
@@ -236,20 +254,12 @@ open class KSOptions {
         nil
     }
 
+    // 虽然只有iOS才支持PIP。但是因为AVSampleBufferDisplayLayer能够支持HDR10+。所以默认还是推荐用AVSampleBufferDisplayLayer
     open func isUseDisplayLayer() -> Bool {
         display == .plane
     }
 
-    private var idetTypeMap = [VideoInterlacingType: Int]()
-    @Published public var videoInterlacingType: VideoInterlacingType?
-    public enum VideoInterlacingType: String {
-        case tff
-        case bff
-        case progressive
-        case undetermined
-    }
-
-    open func io(log: String) {
+    open func urlIO(log: String) {
         if log.starts(with: "Original list of addresses"), dnsStartTime == 0 {
             dnsStartTime = CACurrentMediaTime()
         } else if log.starts(with: "Starting connection attempt to"), tcpStartTime == 0 {
@@ -259,9 +269,10 @@ open class KSOptions {
         }
     }
 
+    private var idetTypeMap = [VideoInterlacingType: UInt]()
     open func filter(log: String) {
-        if log.starts(with: "Repeated Field:") {
-            log.split(separator: ",").forEach { str in
+        if log.starts(with: "Repeated Field:"), autoDeInterlace {
+            for str in log.split(separator: ",") {
                 let map = str.split(separator: ":")
                 if map.count >= 2 {
                     if String(map[0].trimmingCharacters(in: .whitespaces)) == "Multi frame" {
@@ -296,82 +307,102 @@ open class KSOptions {
     }
 
     /**
-            在创建解码器之前可以对KSOptions做一些处理。例如判断fieldOrder为tt或bb的话，那就自动加videofilters
+            在创建解码器之前可以对KSOptions和assetTrack做一些处理。例如判断fieldOrder为tt或bb的话，那就自动加videofilters
      */
-    open func process(assetTrack _: some MediaPlayerTrack) {}
-
-    #if os(tvOS)
-    open func preferredDisplayCriteria(refreshRate _: Float, videoDynamicRange _: Int32) -> AVDisplayCriteria? {
-//        AVDisplayCriteria(refreshRate: refreshRate, videoDynamicRange: videoDynamicRange)
-        nil
-    }
-    #endif
-
-    open func setAudioSession(audioDescriptor: AudioDescriptor) {
-        #if os(macOS)
-        let channels = AVAudioChannelCount(2)
-//        try? AVAudioSession.sharedInstance().setRouteSharingPolicy(.longFormAudio)
-        #else
-        KSOptions.setAudioSession()
-        let isSpatialAudioEnabled: Bool
-        if #available(tvOS 15.0, iOS 15.0, *) {
-            isSpatialAudioEnabled = AVAudioSession.sharedInstance().currentRoute.outputs.contains { $0.isSpatialAudioEnabled }
-            try? AVAudioSession.sharedInstance().setSupportsMultichannelContent(isSpatialAudioEnabled)
-        } else {
-            isSpatialAudioEnabled = false
-        }
-        var channels = audioDescriptor.channels
-        if channels > 2 {
-            let minChannels = min(AVAudioChannelCount(AVAudioSession.sharedInstance().maximumOutputNumberOfChannels), channels)
-            try? AVAudioSession.sharedInstance().setPreferredOutputNumberOfChannels(Int(minChannels))
-            if !isSpatialAudioEnabled {
-                channels = AVAudioChannelCount(AVAudioSession.sharedInstance().preferredOutputNumberOfChannels)
+    open func process(assetTrack: some MediaPlayerTrack) {
+        if assetTrack.mediaType == .video {
+            if [FFmpegFieldOrder.bb, .bt, .tt, .tb].contains(assetTrack.fieldOrder) {
+                // todo 先不要用yadif_videotoolbox，不然会crash。这个后续在看下要怎么解决
+                hardwareDecode = false
+                asynchronousDecompression = false
+                let yadif = hardwareDecode ? "yadif_videotoolbox" : "yadif"
+                var yadifMode = KSOptions.yadifMode
+//                if let assetTrack = assetTrack as? FFmpegAssetTrack {
+//                    if assetTrack.realFrameRate.num == 2 * assetTrack.avgFrameRate.num, assetTrack.realFrameRate.den == assetTrack.avgFrameRate.den {
+//                        if yadifMode == 1 {
+//                            yadifMode = 0
+//                        } else if yadifMode == 3 {
+//                            yadifMode = 2
+//                        }
+//                    }
+//                }
+                videoFilters.append("idet")
+                videoFilters.append("\(yadif)=mode=\(yadifMode):parity=-1:deint=1")
+                if yadifMode == 1 || yadifMode == 3 {
+                    assetTrack.nominalFrameRate = assetTrack.nominalFrameRate * 2
+                }
             }
-        } else {
-            try? AVAudioSession.sharedInstance().setPreferredOutputNumberOfChannels(2)
+        }
+    }
+
+    @MainActor
+    open func updateVideo(refreshRate: Float, isDovi: Bool, formatDescription: CMFormatDescription?) {
+        #if os(tvOS) || os(xrOS)
+        guard let displayManager = UIApplication.shared.windows.first?.avDisplayManager,
+              displayManager.isDisplayCriteriaMatchingEnabled
+        else {
+            return
+        }
+//        快速更改preferredDisplayCriteria，会导致isDisplayModeSwitchInProgress变成true，例如退出一个视频，然后在3s内重新进入的话，
+        if let formatDescription {
+            if KSOptions.displayCriteriaFormatDescriptionEnabled, #available(tvOS 17.0, *) {
+                displayManager.preferredDisplayCriteria = AVDisplayCriteria(refreshRate: refreshRate, formatDescription: formatDescription)
+            } else {
+                let dynamicRange = isDovi ? .dolbyVision : formatDescription.dynamicRange
+                displayManager.preferredDisplayCriteria = AVDisplayCriteria(refreshRate: refreshRate, videoDynamicRange: dynamicRange.rawValue)
+            }
         }
         #endif
-        audioFormat = audioDescriptor.audioFormat(channels: channels)
     }
 
-    open func videoClockSync(main: KSClock, video: KSClock) -> ClockProcessType {
-        var desire = main.getTime() + audioDelay
+    open func videoClockSync(main: KSClock, nextVideoTime: TimeInterval, fps: Float, frameCount: Int) -> (Double, ClockProcessType) {
+        var desire = main.getTime() - videoDelay
         #if !os(macOS)
         desire -= AVAudioSession.sharedInstance().outputLatency
         #endif
-        let diff = video.positionTime + video.duration - desire
-        if diff > 10 || diff < -10 {
-            return .next
-        } else if diff >= 1 / 120 {
+        let diff = nextVideoTime - desire
+//        print("[video] video diff \(diff) nextVideoTime \(nextVideoTime) main \(main.time.seconds)")
+        if diff >= 1 / Double(fps * 2) {
             videoClockDelayCount = 0
-            return .remain
+            return (diff, .remain)
         } else {
-            if diff < -0.04 {
-                KSLog("video delay=\(diff), clock=\(desire), delay count=\(videoClockDelayCount)")
-            }
-            if diff < -0.1 {
+            if diff < -4 / Double(fps) {
                 videoClockDelayCount += 1
-                if diff < -8, videoClockDelayCount % 100 == 0 {
-                    KSLog("video delay seek video track")
-                    return .seek
-                }
-                if diff < -1, videoClockDelayCount % 10 == 0 {
-                    return .flush
-                }
-                if videoClockDelayCount % 2 == 0 {
-                    return .dropNext
+                let log = "[video] video delay=\(diff), clock=\(desire), delay count=\(videoClockDelayCount), frameCount=\(frameCount)"
+                if frameCount == 1 {
+                    if diff < -1, videoClockDelayCount % 10 == 0 {
+                        KSLog("\(log) drop gop Packet")
+                        return (diff, .dropGOPPacket)
+                    } else if videoClockDelayCount % 5 == 0 {
+                        KSLog("\(log) drop next frame")
+                        return (diff, .dropNextFrame)
+                    } else {
+                        return (diff, .next)
+                    }
                 } else {
-                    return .next
+                    if diff < -8, videoClockDelayCount % 100 == 0 {
+                        KSLog("\(log) seek video track")
+                        return (diff, .seek)
+                    }
+                    if diff < -1, videoClockDelayCount % 10 == 0 {
+                        KSLog("\(log) flush video track")
+                        return (diff, .flush)
+                    }
+                    if videoClockDelayCount % 2 == 0 {
+                        KSLog("\(log) drop next frame")
+                        return (diff, .dropNextFrame)
+                    } else {
+                        return (diff, .next)
+                    }
                 }
             } else {
-//                print(CACurrentMediaTime()-video.lastMediaTime)
                 videoClockDelayCount = 0
-                return .next
+                return (diff, .next)
             }
         }
     }
 
-    func availableDynamicRange(_ cotentRange: DynamicRange?) -> DynamicRange? {
+    open func availableDynamicRange(_ cotentRange: DynamicRange?) -> DynamicRange? {
         #if canImport(UIKit)
         let availableHDRModes = AVPlayer.availableHDRModes
         if let preferedDynamicRange = destinationDynamicRange {
@@ -388,9 +419,40 @@ open class KSOptions {
                 return availableHDRModes.dynamicRange
             }
         }
-        #endif
         return cotentRange
+        #else
+        return destinationDynamicRange ?? cotentRange
+        #endif
     }
+
+    open func playerLayerDeinit() {
+        #if os(tvOS) || os(xrOS)
+        runOnMainThread {
+            UIApplication.shared.windows.first?.avDisplayManager.preferredDisplayCriteria = nil
+        }
+        #endif
+    }
+
+    open func liveAdaptivePlaybackRate(loadingState _: LoadingState) -> Float? {
+        nil
+//        if loadingState.isFirst {
+//            return nil
+//        }
+//        if loadingState.loadedTime > preferredForwardBufferDuration + 5 {
+//            return 1.2
+//        } else if loadingState.loadedTime < preferredForwardBufferDuration / 2 {
+//            return 0.8
+//        } else {
+//            return 1
+//        }
+    }
+}
+
+public enum VideoInterlacingType: String {
+    case tff
+    case bff
+    case progressive
+    case undetermined
 }
 
 public extension KSOptions {
@@ -406,12 +468,17 @@ public extension KSOptions {
     static var isAccurateSeek = false
     /// Applies to short videos only
     static var isLoopPlay = false
-    /// 是否自动播放，默认false
-    static var isAutoPlay = false
+    /// 是否自动播放，默认true
+    static var isAutoPlay = true
     /// seek完是否自动播放
     static var isSeekedAutoPlay = true
     static var hardwareDecode = true
+    // 默认不用自研的硬解，因为有些视频的AVPacket的pts顺序是不对的，只有解码后的AVFrame里面的pts是对的。
+    static var asynchronousDecompression = false
     static var isPipPopViewController = false
+    static var canStartPictureInPictureAutomaticallyFromInline = true
+    static var preferredFrame = true
+    static var displayCriteriaFormatDescriptionEnabled = false
     /// 日志级别
     static var logLevel = LogLevel.warning
     static var logger: LogHandler = OSLog(lable: "KSPlayer")
@@ -422,18 +489,64 @@ public extension KSOptions {
         return Int(ncpu)
     }
 
-    internal static func setAudioSession() {
+    static func setAudioSession() {
         #if os(macOS)
 //        try? AVAudioSession.sharedInstance().setRouteSharingPolicy(.longFormAudio)
         #else
         let category = AVAudioSession.sharedInstance().category
-        if category != .playback, category != .playAndRecord {
-            try? AVAudioSession.sharedInstance().setCategory(.playback, mode: .default, policy: .longFormAudio)
+        if category == .playback || category == .playAndRecord {
+            try? AVAudioSession.sharedInstance().setCategory(category, mode: .moviePlayback, policy: .longFormAudio)
+        } else {
+            try? AVAudioSession.sharedInstance().setCategory(.playback, mode: .moviePlayback, policy: .longFormAudio)
         }
-        try? AVAudioSession.sharedInstance().setMode(.moviePlayback)
         try? AVAudioSession.sharedInstance().setActive(true)
         #endif
     }
+
+    #if !os(macOS)
+    static func isSpatialAudioEnabled(channelCount _: AVAudioChannelCount) -> Bool {
+        if #available(tvOS 15.0, iOS 15.0, *) {
+            let isSpatialAudioEnabled = AVAudioSession.sharedInstance().currentRoute.outputs.contains { $0.isSpatialAudioEnabled }
+            try? AVAudioSession.sharedInstance().setSupportsMultichannelContent(isSpatialAudioEnabled)
+            return isSpatialAudioEnabled
+        } else {
+            return false
+        }
+    }
+
+    static func outputNumberOfChannels(channelCount: AVAudioChannelCount) -> AVAudioChannelCount {
+        let maximumOutputNumberOfChannels = AVAudioChannelCount(AVAudioSession.sharedInstance().maximumOutputNumberOfChannels)
+        let preferredOutputNumberOfChannels = AVAudioChannelCount(AVAudioSession.sharedInstance().preferredOutputNumberOfChannels)
+        let isSpatialAudioEnabled = isSpatialAudioEnabled(channelCount: channelCount)
+        let isUseAudioRenderer = KSOptions.audioPlayerType == AudioRendererPlayer.self
+        KSLog("[audio] maximumOutputNumberOfChannels: \(maximumOutputNumberOfChannels), preferredOutputNumberOfChannels: \(preferredOutputNumberOfChannels), isSpatialAudioEnabled: \(isSpatialAudioEnabled), isUseAudioRenderer: \(isUseAudioRenderer) ")
+        let maxRouteChannelsCount = AVAudioSession.sharedInstance().currentRoute.outputs.compactMap {
+            $0.channels?.count
+        }.max() ?? 2
+        KSLog("[audio] currentRoute max channels: \(maxRouteChannelsCount)")
+        var channelCount = channelCount
+        if channelCount > 2 {
+            let minChannels = min(maximumOutputNumberOfChannels, channelCount)
+            #if os(tvOS) || targetEnvironment(simulator)
+            if !(isUseAudioRenderer && isSpatialAudioEnabled) {
+                // 不要用maxRouteChannelsCount来判断，有可能会不准。导致多音道设备也返回2（一开始播放一个2声道，就容易出现），也不能用outputNumberOfChannels来判断，有可能会返回2
+//                channelCount = AVAudioChannelCount(min(AVAudioSession.sharedInstance().outputNumberOfChannels, maxRouteChannelsCount))
+                channelCount = minChannels
+            }
+            #else
+            // iOS 外放是会自动有空间音频功能，但是蓝牙耳机有可能没有空间音频功能或者把空间音频给关了，。所以还是需要处理。
+            if !isSpatialAudioEnabled {
+                channelCount = minChannels
+            }
+            #endif
+        } else {
+            channelCount = 2
+        }
+        // 不在这里设置setPreferredOutputNumberOfChannels,因为这个方法会在获取音轨信息的时候，进行调用。
+        KSLog("[audio] outputNumberOfChannels: \(AVAudioSession.sharedInstance().outputNumberOfChannels) output channelCount: \(channelCount)")
+        return channelCount
+    }
+    #endif
 }
 
 public enum LogLevel: Int32, CustomStringConvertible {
@@ -486,28 +599,31 @@ public extension LogLevel {
 }
 
 public protocol LogHandler {
+    @inlinable
     func log(level: LogLevel, message: CustomStringConvertible, file: String, function: String, line: UInt)
 }
 
 public class OSLog: LogHandler {
-    private let label: String
+    public let label: String
     public init(lable: String) {
         label = lable
     }
 
+    @inlinable
     public func log(level: LogLevel, message: CustomStringConvertible, file: String, function: String, line: UInt) {
         os_log(level.logType, "%@ %@: %@:%d %@ | %@", level.description, label, file, line, function, message.description)
     }
 }
 
 public class FileLog: LogHandler {
-    private let fileHandle: FileHandle
-    private let formatter = DateFormatter()
+    public let fileHandle: FileHandle
+    public let formatter = DateFormatter()
     public init(fileHandle: FileHandle) {
         self.fileHandle = fileHandle
         formatter.dateFormat = "MM-dd HH:mm:ss.SSSSSS"
     }
 
+    @inlinable
     public func log(level: LogLevel, message: CustomStringConvertible, file: String, function: String, line: UInt) {
         let string = String(format: "%@ %@ %@:%d %@ | %@\n", formatter.string(from: Date()), level.description, file, line, function, message.description)
         if let data = string.data(using: .utf8) {
@@ -516,14 +632,21 @@ public class FileLog: LogHandler {
     }
 }
 
-@inlinable public func KSLog(level: LogLevel = KSOptions.logLevel, _ message: @autoclosure () -> CustomStringConvertible, file: String = #file, function: String = #function, line: UInt = #line) {
+@inlinable
+public func KSLog(_ error: @autoclosure () -> Error, file: String = #file, function: String = #function, line: UInt = #line) {
+    KSLog(level: .error, error().localizedDescription, file: file, function: function, line: line)
+}
+
+@inlinable
+public func KSLog(level: LogLevel = .warning, _ message: @autoclosure () -> CustomStringConvertible, file: String = #file, function: String = #function, line: UInt = #line) {
     if level.rawValue <= KSOptions.logLevel.rawValue {
         let fileName = (file as NSString).lastPathComponent
         KSOptions.logger.log(level: level, message: message(), file: fileName, function: function, line: line)
     }
 }
 
-@inlinable public func KSLog(level: LogLevel = .warning, dso: UnsafeRawPointer = #dsohandle, _ message: StaticString, _ args: CVarArg...) {
+@inlinable
+public func KSLog(level: LogLevel = .warning, dso: UnsafeRawPointer = #dsohandle, _ message: StaticString, _ args: CVarArg...) {
     if level.rawValue <= KSOptions.logLevel.rawValue {
         os_log(level.logType, dso: dso, message, args)
     }
@@ -539,16 +662,16 @@ public extension Array {
     }
 }
 
-public class KSClock {
+public struct KSClock {
     public private(set) var lastMediaTime = CACurrentMediaTime()
-    public internal(set) var duration = TimeInterval(0)
-    public internal(set) var positionTime = TimeInterval(0) {
+    public internal(set) var position = Int64(0)
+    public internal(set) var time = CMTime.zero {
         didSet {
             lastMediaTime = CACurrentMediaTime()
         }
     }
 
     func getTime() -> TimeInterval {
-        positionTime + CACurrentMediaTime() - lastMediaTime
+        time.seconds + CACurrentMediaTime() - lastMediaTime
     }
 }
