@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  PlayerDefines.swift
 //  KSPlayer
 //
 //  Created by kintan on 2018/3/9.
@@ -12,11 +12,14 @@ import SwiftUI
 
 #if canImport(UIKit)
 import UIKit
+
 public extension KSOptions {
+    @MainActor
     static var windowScene: UIWindowScene? {
         UIApplication.shared.connectedScenes.first as? UIWindowScene
     }
 
+    @MainActor
     static var sceneSize: CGSize {
         let window = windowScene?.windows.first
         return window?.bounds.size ?? .zero
@@ -25,6 +28,7 @@ public extension KSOptions {
 #else
 import AppKit
 import SwiftUI
+
 public typealias UIView = NSView
 public typealias UIPasteboard = NSPasteboard
 public extension KSOptions {
@@ -39,31 +43,6 @@ public extension KSOptions {
 //        lhs.trackID == rhs.trackID
 //    }
 // }
-
-public extension MediaPlayerTrack {
-    var codecType: FourCharCode {
-        mediaSubType.rawValue
-    }
-
-    func dynamicRange(_ options: KSOptions) -> DynamicRange {
-        let cotentRange: DynamicRange
-        if dovi != nil || codecType.string == "dvhe" || codecType == kCMVideoCodecType_DolbyVisionHEVC {
-            cotentRange = .dolbyVision
-        } else if transferFunction == kCVImageBufferTransferFunction_SMPTE_ST_2084_PQ as String { /// HDR
-            cotentRange = .hdr10
-        } else if transferFunction == kCVImageBufferTransferFunction_ITU_R_2100_HLG as String { /// HDR
-            cotentRange = .hlg
-        } else {
-            cotentRange = .sdr
-        }
-
-        return options.availableDynamicRange(cotentRange) ?? cotentRange
-    }
-
-    var colorSpace: CGColorSpace? {
-        KSOptions.colorSpace(ycbcrMatrix: yCbCrMatrix as CFString?, transferFunction: transferFunction as CFString?)
-    }
-}
 
 public enum DynamicRange: Int32 {
     case sdr = 0
@@ -85,6 +64,21 @@ public enum DynamicRange: Int32 {
         }
     }
     #endif
+}
+
+extension DynamicRange: CustomStringConvertible {
+    public var description: String {
+        switch self {
+        case .sdr:
+            return "SDR"
+        case .hdr10:
+            return "HDR10"
+        case .hlg:
+            return "HDR"
+        case .dolbyVision:
+            return "Dolby Vision"
+        }
+    }
 }
 
 extension DynamicRange {
@@ -147,6 +141,7 @@ public extension FourCharCode {
     }
 }
 
+@MainActor
 public enum DisplayEnum {
     case plane
     // swiftlint:disable identifier_name
@@ -173,7 +168,9 @@ public struct VideoAdaptationState {
 public enum ClockProcessType {
     case remain
     case next
-    case dropNext
+    case dropNextFrame
+    case dropNextPacket
+    case dropGOPPacket
     case flush
     case seek
 }
@@ -186,6 +183,12 @@ public protocol CapacityProtocol {
     var frameMaxCount: Int { get }
     var isEndOfFile: Bool { get }
     var mediaType: AVFoundation.AVMediaType { get }
+}
+
+extension CapacityProtocol {
+    var loadedTime: TimeInterval {
+        TimeInterval(packetCount + frameCount) / TimeInterval(fps)
+    }
 }
 
 public struct LoadingState {
@@ -275,6 +278,12 @@ extension NSError {
         userInfo[NSLocalizedDescriptionKey] = errorCode.description
         self.init(domain: KSPlayerErrorDomain, code: errorCode.rawValue, userInfo: userInfo)
     }
+
+    convenience init(description: String) {
+        var userInfo = [String: Any]()
+        userInfo[NSLocalizedDescriptionKey] = description
+        self.init(domain: KSPlayerErrorDomain, code: 0, userInfo: userInfo)
+    }
 }
 
 extension CMTime {
@@ -325,11 +334,16 @@ func - (left: CGSize, right: CGSize) -> CGSize {
     CGSize(width: left.width - right.width, height: left.height - right.height)
 }
 
-public func runInMainqueue(block: @escaping () -> Void) {
+@inline(__always)
+@preconcurrency
+// @MainActor
+public func runOnMainThread(block: @escaping () -> Void) {
     if Thread.isMainThread {
         block()
     } else {
-        DispatchQueue.main.async(execute: block)
+        Task {
+            await MainActor.run(body: block)
+        }
     }
 }
 
@@ -357,78 +371,125 @@ public extension URL {
     }
 
     func parsePlaylist() async throws -> [(String, URL, [String: String])] {
-        guard let data = try? await data(), let string = String(data: data, encoding: .utf8) else {
-            return []
+        let data = try await data()
+        var entrys = data.parsePlaylist()
+        for i in 0 ..< entrys.count {
+            var entry = entrys[i]
+            if entry.1.path.hasPrefix("./") {
+                entry.1 = deletingLastPathComponent().appendingPathComponent(entry.1.path).standardized
+                entrys[i] = entry
+            }
         }
-        /*
-         #EXTINF:-1 tvg-id="ExampleTV.ua",Example TV (720p) [Not 24/7]
-         #EXTVLCOPT:http-referrer=http://example.com/
-         #EXTVLCOPT:http-user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64)
-         http://example.com/stream.m3u8
-         */
-        return string.components(separatedBy: "#EXTINF:").compactMap { content -> (String, URL, [String: String])? in
-            let content = content.replacingOccurrences(of: "\r\n", with: "\n")
-            let array = content.split(separator: "\n")
-            guard array.count > 1, let last = array.last, let url = URL(string: String(last)) else {
-                return nil
-            }
-            let infos = array[0].split(separator: ",")
-            guard infos.count > 1, let name = infos.last else {
-                return nil
-            }
-            var extinf = [String: String]()
-            let prefix = "#EXTVLCOPT:"
-            for i in 1 ..< (array.count - 1) {
-                let str = array[i]
-                if str.hasPrefix(prefix) {
-                    let keyValue = str.dropFirst(prefix.count).split(separator: "=")
-                    if keyValue.count == 2 {
-                        extinf[String(keyValue[0])] = String(keyValue[1])
-                    }
-                }
-            }
-            let tvgString: Substring
-            if infos.count > 2 {
-                extinf["duration"] = String(infos[0])
-                tvgString = infos[1]
-            } else {
-                tvgString = infos[0]
-            }
-            tvgString.split(separator: " ").forEach { str in
-                let keyValue = str.split(separator: "=")
-                if keyValue.count == 2 {
-                    extinf[String(keyValue[0])] = keyValue[1].trimmingCharacters(in: CharacterSet(charactersIn: #"""#))
-                } else {
-                    extinf["duration"] = String(keyValue[0])
-                }
-            }
-            return (String(name), url, extinf)
-        }
+        return entrys
     }
 
-    func data() async throws -> Data {
+    func data(userAgent: String? = nil) async throws -> Data {
         if isFileURL {
             return try Data(contentsOf: self)
         } else {
-            let (data, _) = try await URLSession.shared.data(from: self)
+            var request = URLRequest(url: self)
+            if let userAgent {
+                request.addValue(userAgent, forHTTPHeaderField: "User-Agent")
+            }
+            let (data, _) = try await URLSession.shared.data(for: request)
             return data
         }
     }
 
-    func download(completion: @escaping ((String, URL) -> Void)) {
-        URLSession.shared.downloadTask(with: self) { url, response, _ in
-            guard let url, let response = response as? HTTPURLResponse else {
+    func download(userAgent: String? = nil, completion: @escaping ((String, URL) -> Void)) {
+        var request = URLRequest(url: self)
+        if let userAgent {
+            request.addValue(userAgent, forHTTPHeaderField: "User-Agent")
+        }
+        let task = URLSession.shared.downloadTask(with: request) { url, response, _ in
+            guard let url, let response else {
                 return
             }
-            let httpFileName = "attachment; filename="
-            var filename = url.lastPathComponent
-            if var disposition = response.value(forHTTPHeaderField: "Content-Disposition"), disposition.hasPrefix(httpFileName) {
-                disposition.removeFirst(httpFileName.count)
-                filename = disposition
-            }
             // 下载的临时文件要马上就用。不然可能会马上被清空
-            completion(filename, url)
-        }.resume()
+            completion(response.suggestedFilename ?? url.lastPathComponent, url)
+        }
+        task.resume()
+    }
+}
+
+public extension Data {
+    func parsePlaylist() -> [(String, URL, [String: String])] {
+        guard let string = String(data: self, encoding: .utf8) else {
+            return []
+        }
+        let scanner = Scanner(string: string)
+        var entrys = [(String, URL, [String: String])]()
+        guard let symbol = scanner.scanUpToCharacters(from: .newlines), symbol.contains("#EXTM3U") else {
+            return []
+        }
+        while !scanner.isAtEnd {
+            if let entry = scanner.parseM3U() {
+                entrys.append(entry)
+            }
+        }
+        return entrys
+    }
+}
+
+extension Scanner {
+    /*
+     #EXTINF:-1 tvg-id="ExampleTV.ua" tvg-logo="https://image.com" group-title="test test", Example TV (720p) [Not 24/7]
+     #EXTVLCOPT:http-referrer=http://example.com/
+     #EXTVLCOPT:http-user-agent=Mozilla/5.0 (Windows NT 10.0; Win64; x64)
+     http://example.com/stream.m3u8
+     */
+    func parseM3U() -> (String, URL, [String: String])? {
+        if scanString("#EXTINF:") == nil {
+            _ = scanUpToCharacters(from: .newlines)
+            return nil
+        }
+        var extinf = [String: String]()
+        if let duration = scanDouble() {
+            extinf["duration"] = String(duration)
+        }
+        while scanString(",") == nil {
+            let key = scanUpToString("=")
+            _ = scanString("=\"")
+            let value = scanUpToString("\"")
+            _ = scanString("\"")
+            if let key, let value {
+                extinf[key] = value
+            }
+        }
+        let title = scanUpToCharacters(from: .newlines)
+        while scanString("#EXT") != nil {
+            if scanString("VLCOPT:") != nil {
+                let key = scanUpToString("=")
+                _ = scanString("=")
+                let value = scanUpToCharacters(from: .newlines)
+                if let key, let value {
+                    extinf[key] = value
+                }
+            } else {
+                let key = scanUpToString(":")
+                _ = scanString(":")
+                let value = scanUpToCharacters(from: .newlines)
+                if let key, let value {
+                    extinf[key] = value
+                }
+            }
+        }
+        let urlString = scanUpToCharacters(from: .newlines)
+        if let urlString, let url = URL(string: urlString) {
+            return (title ?? url.lastPathComponent, url, extinf)
+        }
+        return nil
+    }
+}
+
+extension HTTPURLResponse {
+    var filename: String? {
+        let httpFileName = "attachment; filename="
+        if var disposition = value(forHTTPHeaderField: "Content-Disposition"), disposition.hasPrefix(httpFileName) {
+            disposition.removeFirst(httpFileName.count)
+            return disposition
+        }
+        return nil
     }
 }
 
@@ -489,6 +550,27 @@ public extension Int {
     }
 }
 
+public extension FixedWidthInteger {
+    var kmFormatted: String {
+        Double(self).kmFormatted
+    }
+}
+
+public extension Double {
+    var kmFormatted: String {
+//        return .formatted(.number.notation(.compactName))
+        if self >= 1_000_000 {
+            return String(format: "%.1fM", locale: Locale.current, self / 1_000_000)
+//                .replacingOccurrences(of: ".0", with: "")
+        } else if self >= 10000, self <= 999_999 {
+            return String(format: "%.1fK", locale: Locale.current, self / 1000)
+//                .replacingOccurrences(of: ".0", with: "")
+        } else {
+            return String(format: "%.0f", locale: Locale.current, self)
+        }
+    }
+}
+
 extension TextAlignment: RawRepresentable {
     public typealias RawValue = String
     public init?(rawValue: RawValue) {
@@ -516,6 +598,38 @@ extension TextAlignment: RawRepresentable {
 }
 
 extension TextAlignment: Identifiable {
+    public var id: Self { self }
+}
+
+extension HorizontalAlignment: Hashable, RawRepresentable {
+    public typealias RawValue = String
+    public init?(rawValue: RawValue) {
+        if rawValue == "Leading" {
+            self = .leading
+        } else if rawValue == "Center" {
+            self = .center
+        } else if rawValue == "Trailing" {
+            self = .trailing
+        } else {
+            return nil
+        }
+    }
+
+    public var rawValue: RawValue {
+        switch self {
+        case .leading:
+            return "Leading"
+        case .center:
+            return "Center"
+        case .trailing:
+            return "Trailing"
+        default:
+            return ""
+        }
+    }
+}
+
+extension HorizontalAlignment: Identifiable {
     public var id: Self { self }
 }
 
@@ -560,7 +674,7 @@ extension Color: RawRepresentable {
         }
 
         do {
-            let color = try NSKeyedUnarchiver.unarchiveTopLevelObjectWithData(data) as? UIColor ?? .black
+            let color = try NSKeyedUnarchiver.unarchivedObject(ofClass: UIColor.self, from: data) ?? .black
             self = Color(color)
         } catch {
             self = .black
