@@ -1,5 +1,5 @@
 //
-//  PlayerController.swift
+//  PlayerView.swift
 //  VoiceNote
 //
 //  Created by kintan on 2018/8/16.
@@ -33,8 +33,9 @@ public protocol PlayerControllerDelegate: AnyObject {
     func playerController(finish error: Error?)
     func playerController(maskShow: Bool)
     func playerController(action: PlayerButtonType)
-    // bufferedCount: 0代表首次加载
+    // `bufferedCount: 0` indicates first time loading
     func playerController(bufferedCount: Int, consumeTime: TimeInterval)
+    func playerController(seek: TimeInterval)
 }
 
 open class PlayerView: UIView, KSPlayerLayerDelegate, KSSliderDelegate {
@@ -48,7 +49,7 @@ open class PlayerView: UIView, KSPlayerLayerDelegate, KSSliderDelegate {
     public weak var delegate: ControllerDelegate?
     public let toolBar = PlayerToolBar()
     public let srtControl = SubtitleModel()
-    // Closure fired when play time changed
+    // Listen to play time change
     public var playTimeDidChange: ((TimeInterval, TimeInterval) -> Void)?
     public var backBlock: (() -> Void)?
     public convenience init() {
@@ -130,6 +131,7 @@ open class PlayerView: UIView, KSPlayerLayerDelegate, KSSliderDelegate {
     open func play() {
         becomeFirstResponder()
         playerLayer?.play()
+        toolBar.playButton.isSelected = true
     }
 
     open func pause() {
@@ -158,7 +160,8 @@ open class PlayerView: UIView, KSPlayerLayerDelegate, KSSliderDelegate {
         if event == .valueChanged {
             toolBar.currentTime = value
         } else if event == .touchUpInside {
-            seek(time: value) { _ in
+            seek(time: value) { [weak self] _ in
+                self?.delegate?.playerController(seek: value)
             }
         }
     }
@@ -170,10 +173,9 @@ open class PlayerView: UIView, KSPlayerLayerDelegate, KSSliderDelegate {
         if state == .readyToPlay {
             totalTime = layer.player.duration
             toolBar.isSeekable = layer.player.seekable
+            toolBar.playButton.isSelected = true
         } else if state == .playedToTheEnd || state == .paused || state == .error {
             toolBar.playButton.isSelected = false
-        } else if state.isPlaying {
-            toolBar.playButton.isSelected = true
         }
     }
 
