@@ -1,5 +1,5 @@
 //
-//  File.swift
+//  Utility.swift
 //  KSPlayer
 //
 //  Created by kintan on 2018/3/9.
@@ -76,7 +76,31 @@ extension String {
 }
 
 public extension UIColor {
-    convenience init(hex: Int, alpha: CGFloat = 1) {
+    convenience init?(assColor: String) {
+        var colorString = assColor
+        // 移除颜色字符串中的前缀 &H 和后缀 &
+        if colorString.hasPrefix("&H") {
+            colorString = String(colorString.dropFirst(2))
+        }
+        if colorString.hasSuffix("&") {
+            colorString = String(colorString.dropLast())
+        }
+        if let hex = Scanner(string: colorString).scanInt(representation: .hexadecimal) {
+            self.init(abgr: hex)
+        } else {
+            return nil
+        }
+    }
+
+    convenience init(abgr hex: Int) {
+        let alpha = 1 - (CGFloat(hex >> 24 & 0xFF) / 255)
+        let blue = CGFloat((hex >> 16) & 0xFF)
+        let green = CGFloat((hex >> 8) & 0xFF)
+        let red = CGFloat(hex & 0xFF)
+        self.init(red: red / 255.0, green: green / 255.0, blue: blue / 255.0, alpha: alpha)
+    }
+
+    convenience init(rgb hex: Int, alpha: CGFloat = 1) {
         let red = CGFloat((hex >> 16) & 0xFF)
         let green = CGFloat((hex >> 8) & 0xFF)
         let blue = CGFloat(hex & 0xFF)
