@@ -3,6 +3,7 @@ import FFmpegKit
 import Libavcodec
 import Libavfilter
 import Libavformat
+
 func toDictionary(_ native: OpaquePointer?) -> [String: String] {
     var dict = [String: String]()
     if let native {
@@ -28,20 +29,22 @@ extension UnsafeMutablePointer where Pointee == AVCodecContext {
                     if deviceCtx == nil {
                         break
                     }
-                    var framesCtx = av_hwframe_ctx_alloc(deviceCtx)
-                    if let framesCtx {
-                        let framesCtxData = UnsafeMutableRawPointer(framesCtx.pointee.data)
-                            .bindMemory(to: AVHWFramesContext.self, capacity: 1)
-                        framesCtxData.pointee.format = AV_PIX_FMT_VIDEOTOOLBOX
-                        framesCtxData.pointee.sw_format = ctx.pointee.pix_fmt.bestPixelFormat()
-                        framesCtxData.pointee.width = ctx.pointee.width
-                        framesCtxData.pointee.height = ctx.pointee.height
-                    }
-                    if av_hwframe_ctx_init(framesCtx) != 0 {
-                        av_buffer_unref(&framesCtx)
-                        break
-                    }
-                    ctx.pointee.hw_frames_ctx = framesCtx
+                    // 只要有hw_device_ctx就可以了。不需要hw_frames_ctx
+                    ctx.pointee.hw_device_ctx = deviceCtx
+//                    var framesCtx = av_hwframe_ctx_alloc(deviceCtx)
+//                    if let framesCtx {
+//                        let framesCtxData = UnsafeMutableRawPointer(framesCtx.pointee.data)
+//                            .bindMemory(to: AVHWFramesContext.self, capacity: 1)
+//                        framesCtxData.pointee.format = AV_PIX_FMT_VIDEOTOOLBOX
+//                        framesCtxData.pointee.sw_format = ctx.pointee.pix_fmt.bestPixelFormat
+//                        framesCtxData.pointee.width = ctx.pointee.width
+//                        framesCtxData.pointee.height = ctx.pointee.height
+//                    }
+//                    if av_hwframe_ctx_init(framesCtx) != 0 {
+//                        av_buffer_unref(&framesCtx)
+//                        break
+//                    }
+//                    ctx.pointee.hw_frames_ctx = framesCtx
                     return fmt[i]
                 }
                 i += 1
@@ -74,7 +77,7 @@ extension AVCodecContext {
 }
 
 extension AVCodecParameters {
-    mutating func ceateContext(options: KSOptions) throws -> UnsafeMutablePointer<AVCodecContext> {
+    mutating func createContext(options: KSOptions?) throws -> UnsafeMutablePointer<AVCodecContext> {
         var codecContextOption = avcodec_alloc_context3(nil)
         guard let codecContext = codecContextOption else {
             throw NSError(errorCode: .codecContextCreate)
@@ -84,7 +87,7 @@ extension AVCodecParameters {
             avcodec_free_context(&codecContextOption)
             throw NSError(errorCode: .codecContextSetParam, avErrorCode: result)
         }
-        if codec_type == AVMEDIA_TYPE_VIDEO, options.hardwareDecode {
+        if codec_type == AVMEDIA_TYPE_VIDEO, options?.hardwareDecode ?? false {
             codecContext.getFormat()
         }
         guard let codec = avcodec_find_decoder(codecContext.pointee.codec_id) else {
@@ -93,16 +96,22 @@ extension AVCodecParameters {
         }
         codecContext.pointee.codec_id = codec.pointee.id
         codecContext.pointee.flags2 |= AV_CODEC_FLAG2_FAST
-        var lowres = options.lowres
-        if lowres > codec.pointee.max_lowres {
-            lowres = codec.pointee.max_lowres
+        if options?.codecLowDelay == true {
+            codecContext.pointee.flags |= AV_CODEC_FLAG_LOW_DELAY
         }
-        codecContext.pointee.lowres = Int32(lowres)
-        var avOptions = options.decoderOptions.avOptions
-        if lowres > 0 {
-            av_dict_set_int(&avOptions, "lowres", Int64(lowres), 0)
+        var avOptions = options?.decoderOptions.avOptions
+        if let options {
+            var lowres = options.lowres
+            if lowres > codec.pointee.max_lowres {
+                lowres = codec.pointee.max_lowres
+            }
+            codecContext.pointee.lowres = Int32(lowres)
+            if lowres > 0 {
+                av_dict_set_int(&avOptions, "lowres", Int64(lowres), 0)
+            }
         }
         result = avcodec_open2(codecContext, codec, &avOptions)
+        av_dict_free(&avOptions)
         guard result == 0 else {
             avcodec_free_context(&codecContextOption)
             throw NSError(errorCode: .codesContextOpen, avErrorCode: result)
@@ -218,12 +227,12 @@ extension AVChromaLocation {
 }
 
 extension AVPixelFormat {
-    func bitDepth() -> Int32 {
+    var bitDepth: Int32 {
         let descriptor = av_pix_fmt_desc_get(self)
         return descriptor?.pointee.comp.0.depth ?? 8
     }
 
-    func planeCount() -> UInt8 {
+    var planeCount: UInt8 {
         if let desc = av_pix_fmt_desc_get(self) {
             switch desc.pointee.nb_components {
             case 3:
@@ -238,7 +247,16 @@ extension AVPixelFormat {
         }
     }
 
-    func bestPixelFormat() -> AVPixelFormat {
+    var leftShift: UInt8 {
+        if [AV_PIX_FMT_YUV420P10LE, AV_PIX_FMT_YUV422P10LE, AV_PIX_FMT_YUV444P10LE].contains(self) {
+            return 6
+        } else {
+            return 0
+        }
+    }
+
+    // videotoolbox_best_pixel_format
+    var bestPixelFormat: AVPixelFormat {
         if let desc = av_pix_fmt_desc_get(self) {
             if desc.pointee.flags & UInt64(AV_PIX_FMT_FLAG_ALPHA) != 0 {
                 return AV_PIX_FMT_AYUV64LE
@@ -253,7 +271,7 @@ extension AVPixelFormat {
             if desc.pointee.log2_chroma_h == 0 {
                 return depth <= 8 ? AV_PIX_FMT_NV16 : AV_PIX_FMT_P210LE
             }
-            return depth > 8 ? AV_PIX_FMT_P010LE : AV_PIX_FMT_NV12
+            return depth <= 8 ? AV_PIX_FMT_NV12 : AV_PIX_FMT_P010LE
         } else {
             return AV_PIX_FMT_NV12
         }
@@ -264,6 +282,7 @@ extension AVPixelFormat {
     func osType(fullRange: Bool = false) -> OSType? {
         switch self {
         case AV_PIX_FMT_MONOBLACK: return kCVPixelFormatType_1Monochrome
+//        case AV_PIX_FMT_PAL8: return kCVPixelFormatType_32RGBA
         case AV_PIX_FMT_GRAY8: return kCVPixelFormatType_OneComponent8
         case AV_PIX_FMT_RGB555BE: return kCVPixelFormatType_16BE555
         case AV_PIX_FMT_RGB555LE: return kCVPixelFormatType_16LE555
@@ -280,23 +299,19 @@ extension AVPixelFormat {
         case AV_PIX_FMT_RGBA: return kCVPixelFormatType_32RGBA
         case AV_PIX_FMT_BGR48BE, AV_PIX_FMT_BGR48LE: return kCVPixelFormatType_48RGB
         case AV_PIX_FMT_NV12: return fullRange ? kCVPixelFormatType_420YpCbCr8BiPlanarFullRange : kCVPixelFormatType_420YpCbCr8BiPlanarVideoRange
-        case AV_PIX_FMT_P010BE, AV_PIX_FMT_P010LE: return fullRange ? kCVPixelFormatType_420YpCbCr10BiPlanarFullRange : kCVPixelFormatType_420YpCbCr10BiPlanarVideoRange
-        case AV_PIX_FMT_YUV420P10BE, AV_PIX_FMT_YUV420P10LE: return kCVPixelFormatType_420YpCbCr10BiPlanarVideoRange
-        case AV_PIX_FMT_YUV420P: return kCVPixelFormatType_420YpCbCr8Planar
-        //  AVSampleBufferDisplayLayer不能显示 kCVPixelFormatType_420YpCbCr8PlanarFullRange
-        case AV_PIX_FMT_YUVJ420P: return kCVPixelFormatType_420YpCbCr8BiPlanarFullRange
-        case AV_PIX_FMT_NV16: return fullRange ? kCVPixelFormatType_422YpCbCr8BiPlanarFullRange : kCVPixelFormatType_422YpCbCr8BiPlanarVideoRange
+        //  AVSampleBufferDisplayLayer不能显示 kCVPixelFormatType_420YpCbCr8PlanarFullRange,所以换成是kCVPixelFormatType_420YpCbCr8BiPlanarFullRange
+        case AV_PIX_FMT_YUV420P, AV_PIX_FMT_YUVJ420P: return fullRange ? kCVPixelFormatType_420YpCbCr8BiPlanarFullRange : kCVPixelFormatType_420YpCbCr8Planar
+        case AV_PIX_FMT_P010BE, AV_PIX_FMT_P010LE, AV_PIX_FMT_YUV420P10BE, AV_PIX_FMT_YUV420P10LE: return fullRange ? kCVPixelFormatType_420YpCbCr10BiPlanarFullRange : kCVPixelFormatType_420YpCbCr10BiPlanarVideoRange
         case AV_PIX_FMT_UYVY422: return kCVPixelFormatType_422YpCbCr8
         case AV_PIX_FMT_YUYV422: return kCVPixelFormatType_422YpCbCr8_yuvs
-        case AV_PIX_FMT_P210BE, AV_PIX_FMT_P210LE: return fullRange ? kCVPixelFormatType_422YpCbCr10BiPlanarFullRange : kCVPixelFormatType_422YpCbCr10BiPlanarVideoRange
-        case AV_PIX_FMT_YUV422P10BE, AV_PIX_FMT_YUV422P10LE: return kCVPixelFormatType_422YpCbCr10
-        case AV_PIX_FMT_P216BE, AV_PIX_FMT_P216LE: return kCVPixelFormatType_422YpCbCr16BiPlanarVideoRange
-        case AV_PIX_FMT_YUV422P16BE, AV_PIX_FMT_YUV422P16LE: return kCVPixelFormatType_422YpCbCr16
-        case AV_PIX_FMT_NV24: return fullRange ? kCVPixelFormatType_444YpCbCr8BiPlanarFullRange : kCVPixelFormatType_444YpCbCr8BiPlanarVideoRange
-        case AV_PIX_FMT_YUV444P: return kCVPixelFormatType_444YpCbCr8
+        case AV_PIX_FMT_NV16: return fullRange ? kCVPixelFormatType_422YpCbCr8BiPlanarFullRange : kCVPixelFormatType_422YpCbCr8BiPlanarVideoRange
+        case AV_PIX_FMT_YUV422P, AV_PIX_FMT_YUVJ422P: return fullRange ? kCVPixelFormatType_422YpCbCr8BiPlanarFullRange : kCVPixelFormatType_422YpCbCr8BiPlanarVideoRange
+        case AV_PIX_FMT_Y210BE, AV_PIX_FMT_Y210LE: return kCVPixelFormatType_422YpCbCr10
+        case AV_PIX_FMT_P210BE, AV_PIX_FMT_P210LE, AV_PIX_FMT_YUV422P10BE, AV_PIX_FMT_YUV422P10LE: return fullRange ? kCVPixelFormatType_422YpCbCr10BiPlanarFullRange : kCVPixelFormatType_422YpCbCr10BiPlanarVideoRange
+        case AV_PIX_FMT_P216BE, AV_PIX_FMT_P216LE, AV_PIX_FMT_YUV422P16BE, AV_PIX_FMT_YUV422P16LE: return kCVPixelFormatType_422YpCbCr16BiPlanarVideoRange
+        case AV_PIX_FMT_NV24, AV_PIX_FMT_YUV444P: return fullRange ? kCVPixelFormatType_444YpCbCr8BiPlanarFullRange : kCVPixelFormatType_444YpCbCr8BiPlanarVideoRange
         case AV_PIX_FMT_YUVA444P: return kCVPixelFormatType_4444YpCbCrA8R
-        case AV_PIX_FMT_P410BE, AV_PIX_FMT_P410LE: return fullRange ? kCVPixelFormatType_444YpCbCr10BiPlanarFullRange : kCVPixelFormatType_444YpCbCr10BiPlanarVideoRange
-        case AV_PIX_FMT_YUV444P10BE, AV_PIX_FMT_YUV444P10LE: return kCVPixelFormatType_444YpCbCr10
+        case AV_PIX_FMT_P410BE, AV_PIX_FMT_P410LE, AV_PIX_FMT_YUV444P10BE, AV_PIX_FMT_YUV444P10LE: return fullRange ? kCVPixelFormatType_444YpCbCr10BiPlanarFullRange : kCVPixelFormatType_444YpCbCr10BiPlanarVideoRange
         case AV_PIX_FMT_P416BE, AV_PIX_FMT_P416LE: return kCVPixelFormatType_444YpCbCr16BiPlanarVideoRange
         case AV_PIX_FMT_AYUV64BE, AV_PIX_FMT_AYUV64LE: return kCVPixelFormatType_4444AYpCbCr16
         case AV_PIX_FMT_YUVA444P16BE, AV_PIX_FMT_YUVA444P16LE: return kCVPixelFormatType_4444AYpCbCr16
@@ -368,9 +383,7 @@ extension AVRational {
 
 extension AVBufferSrcParameters: Equatable {
     public static func == (lhs: AVBufferSrcParameters, rhs: AVBufferSrcParameters) -> Bool {
-        lhs.format == rhs.format && lhs.time_base == rhs.time_base &&
-            lhs.width == rhs.width && lhs.height == rhs.height && lhs.sample_aspect_ratio == rhs.sample_aspect_ratio &&
-            lhs.sample_rate == rhs.sample_rate && lhs.ch_layout == rhs.ch_layout
+        lhs.format == rhs.format && lhs.width == rhs.width && lhs.height == rhs.height && lhs.sample_aspect_ratio == rhs.sample_aspect_ratio && lhs.sample_rate == rhs.sample_rate && lhs.ch_layout == rhs.ch_layout
     }
 
     var arg: String {
@@ -393,16 +406,16 @@ extension AVChannelLayout: Equatable {
 
 extension AVChannelLayout: CustomStringConvertible {
     static let defaultValue = AVChannelLayout(order: AV_CHANNEL_ORDER_NATIVE, nb_channels: 2, u: AVChannelLayout.__Unnamed_union_u(mask: swift_AV_CH_LAYOUT_STEREO), opaque: nil)
-    var layoutTag: AudioChannelLayoutTag {
-        KSLog("FFmepg channelLayout: \(self) order: \(order)")
+    var layoutTag: AudioChannelLayoutTag? {
+        KSLog("[audio] FFmepg AVChannelLayout: \(self) order: \(order) mask: \(u.mask)")
         let tag = layoutMapTuple.first { _, mask in
             u.mask == mask
         }?.tag
         if let tag {
             return tag
         } else {
-            KSLog("can not find AudioChannelLayoutTag FFmepg channelLayout: \(self) order: \(order)")
-            return kAudioChannelLayoutTag_Stereo
+            KSLog("[audio] can not find AudioChannelLayoutTag FFmepg channelLayout: \(self) order: \(order) mask: \(u.mask)")
+            return nil
         }
     }
 
