@@ -1,5 +1,5 @@
 //
-//  CacheDataSouce.swift
+//  EmbedDataSouce.swift
 //  KSPlayer-7de52535
 //
 //  Created by kintan on 2018/8/7.
@@ -9,28 +9,16 @@ import Libavcodec
 import Libavutil
 
 extension FFmpegAssetTrack: SubtitleInfo {
-    public func subtitle(isEnabled: Bool) {
-        if isImageSubtitle {
-            self.isEnabled = isEnabled
-        }
-    }
-
     public var subtitleID: String {
         String(trackID)
     }
 }
 
 extension FFmpegAssetTrack: KSSubtitleProtocol {
-    public func search(for time: TimeInterval) -> SubtitlePart? {
-        if isImageSubtitle {
-            return subtitle?.outputRenderQueue.pop { item -> Bool in
-                item.part < time || item.part == time
-            }?.part
-        } else {
-            return subtitle?.outputRenderQueue.search { item -> Bool in
-                item.part == time
-            }?.part
-        }
+    public func search(for time: TimeInterval) -> [SubtitlePart] {
+        subtitle?.outputRenderQueue.search { item -> Bool in
+            item.part < time || item.part == time
+        }.map(\.part) ?? []
     }
 }
 
