@@ -8,20 +8,19 @@
 import Foundation
 import Libavfilter
 import Libavutil
+
 class MEFilter {
     private var graph: UnsafeMutablePointer<AVFilterGraph>?
-    private var bufferContext: UnsafeMutablePointer<AVFilterContext>?
+    private var bufferSrcContext: UnsafeMutablePointer<AVFilterContext>?
     private var bufferSinkContext: UnsafeMutablePointer<AVFilterContext>?
-    private var outputFrame = av_frame_alloc()
     private var filters: String?
-    private let timebase: Timebase
+    let timebase: Timebase
     private let isAudio: Bool
     private var params = AVBufferSrcParameters()
     private let nominalFrameRate: Float
     deinit {
         graph?.pointee.opaque = nil
         avfilter_graph_free(&graph)
-        av_frame_free(&outputFrame)
     }
 
     public init(timebase: Timebase, isAudio: Bool, nominalFrameRate: Float, options: KSOptions) {
@@ -35,12 +34,10 @@ class MEFilter {
     private func setup(filters: String) -> Bool {
         var inputs = avfilter_inout_alloc()
         var outputs = avfilter_inout_alloc()
-        defer {
-            avfilter_inout_free(&inputs)
-            avfilter_inout_free(&outputs)
-        }
         var ret = avfilter_graph_parse2(graph, filters, &inputs, &outputs)
         guard ret >= 0, let graph, let inputs, let outputs else {
+            avfilter_inout_free(&inputs)
+            avfilter_inout_free(&outputs)
             return false
         }
         let bufferSink = avfilter_get_by_name(isAudio ? "abuffersink" : "buffersink")
@@ -49,26 +46,63 @@ class MEFilter {
         ret = avfilter_link(outputs.pointee.filter_ctx, UInt32(outputs.pointee.pad_idx), bufferSinkContext, 0)
         guard ret >= 0 else { return false }
         let buffer = avfilter_get_by_name(isAudio ? "abuffer" : "buffer")
-        bufferContext = avfilter_graph_alloc_filter(graph, buffer, "in")
-        guard bufferContext != nil else { return false }
-        av_buffersrc_parameters_set(bufferContext, &params)
-        ret = avfilter_init_str(bufferContext, nil)
+        bufferSrcContext = avfilter_graph_alloc_filter(graph, buffer, "in")
+        guard bufferSrcContext != nil else { return false }
+        av_buffersrc_parameters_set(bufferSrcContext, &params)
+        ret = avfilter_init_str(bufferSrcContext, nil)
         guard ret >= 0 else { return false }
-        ret = avfilter_link(bufferContext, 0, inputs.pointee.filter_ctx, UInt32(inputs.pointee.pad_idx))
+        ret = avfilter_link(bufferSrcContext, 0, inputs.pointee.filter_ctx, UInt32(inputs.pointee.pad_idx))
         guard ret >= 0 else { return false }
-        var hwDeviceCtx: UnsafeMutablePointer<AVBufferRef>?
-        av_hwdevice_ctx_create(&hwDeviceCtx, AV_HWDEVICE_TYPE_VIDEOTOOLBOX, nil, nil, 0)
-        if let hwDeviceCtx {
-            (0 ..< graph.pointee.nb_filters).forEach { i in
-                graph.pointee.filters[Int(i)]?.pointee.hw_device_ctx = av_buffer_ref(hwDeviceCtx)
-            }
+        if let ctx = params.hw_frames_ctx {
+            let framesCtxData = UnsafeMutableRawPointer(ctx.pointee.data).bindMemory(to: AVHWFramesContext.self, capacity: 1)
+            inputs.pointee.filter_ctx.pointee.hw_device_ctx = framesCtxData.pointee.device_ref
+//                    outputs.pointee.filter_ctx.pointee.hw_device_ctx = framesCtxData.pointee.device_ref
+//                    bufferSrcContext?.pointee.hw_device_ctx = framesCtxData.pointee.device_ref
+//                    bufferSinkContext?.pointee.hw_device_ctx = framesCtxData.pointee.device_ref
         }
         ret = avfilter_graph_config(graph, nil)
         guard ret >= 0 else { return false }
         return true
     }
 
-    public func filter(options: KSOptions, inputFrame: UnsafeMutablePointer<AVFrame>, hwFramesCtx: UnsafeMutablePointer<AVBufferRef>?) -> UnsafeMutablePointer<AVFrame> {
+    private func setup2(filters: String) -> Bool {
+        guard let graph else {
+            return false
+        }
+        let bufferName = isAudio ? "abuffer" : "buffer"
+        let bufferSrc = avfilter_get_by_name(bufferName)
+        var ret = avfilter_graph_create_filter(&bufferSrcContext, bufferSrc, "ksplayer_\(bufferName)", params.arg, nil, graph)
+        av_buffersrc_parameters_set(bufferSrcContext, &params)
+        let bufferSink = avfilter_get_by_name(bufferName + "sink")
+        ret = avfilter_graph_create_filter(&bufferSinkContext, bufferSink, "ksplayer_\(bufferName)sink", nil, nil, graph)
+        guard ret >= 0 else { return false }
+        //        av_opt_set_int_list(bufferSinkContext, "pix_fmts", [AV_PIX_FMT_GRAY8, AV_PIX_FMT_NONE] AV_PIX_FMT_NONE,AV_OPT_SEARCH_CHILDREN)
+        var inputs = avfilter_inout_alloc()
+        var outputs = avfilter_inout_alloc()
+        outputs?.pointee.name = strdup("in")
+        outputs?.pointee.filter_ctx = bufferSrcContext
+        outputs?.pointee.pad_idx = 0
+        outputs?.pointee.next = nil
+        inputs?.pointee.name = strdup("out")
+        inputs?.pointee.filter_ctx = bufferSinkContext
+        inputs?.pointee.pad_idx = 0
+        inputs?.pointee.next = nil
+        let filterNb = Int(graph.pointee.nb_filters)
+        ret = avfilter_graph_parse_ptr(graph, filters, &inputs, &outputs, nil)
+        guard ret >= 0 else {
+            avfilter_inout_free(&inputs)
+            avfilter_inout_free(&outputs)
+            return false
+        }
+        for i in 0 ..< Int(graph.pointee.nb_filters) - filterNb {
+            swap(&graph.pointee.filters[i], &graph.pointee.filters[i + filterNb])
+        }
+        ret = avfilter_graph_config(graph, nil)
+        guard ret >= 0 else { return false }
+        return true
+    }
+
+    public func filter(options: KSOptions, inputFrame: UnsafeMutablePointer<AVFrame>, completionHandler: (UnsafeMutablePointer<AVFrame>) -> Void) {
         let filters: String
         if isAudio {
             filters = options.audioFilters.joined(separator: ",")
@@ -78,8 +112,9 @@ class MEFilter {
             }
             filters = options.videoFilters.joined(separator: ",")
         }
-        guard filters.count > 0 else {
-            return inputFrame
+        guard !filters.isEmpty else {
+            completionHandler(inputFrame)
+            return
         }
         var params = AVBufferSrcParameters()
         params.format = inputFrame.pointee.format
@@ -88,24 +123,32 @@ class MEFilter {
         params.height = inputFrame.pointee.height
         params.sample_aspect_ratio = inputFrame.pointee.sample_aspect_ratio
         params.frame_rate = AVRational(num: 1, den: Int32(nominalFrameRate))
-        params.hw_frames_ctx = hwFramesCtx
+        if let ctx = inputFrame.pointee.hw_frames_ctx {
+            params.hw_frames_ctx = av_buffer_ref(ctx)
+        }
         params.sample_rate = inputFrame.pointee.sample_rate
         params.ch_layout = inputFrame.pointee.ch_layout
         if self.params != params || self.filters != filters {
             self.params = params
             self.filters = filters
             if !setup(filters: filters) {
-                return inputFrame
+                completionHandler(inputFrame)
+                return
             }
         }
         if graph?.pointee.sink_links_count == 0 {
-            return inputFrame
+            completionHandler(inputFrame)
+            return
         }
-        var ret = av_buffersrc_add_frame_flags(bufferContext, inputFrame, Int32(AV_BUFFERSRC_FLAG_KEEP_REF))
-        guard ret == 0 else { return inputFrame }
-        av_frame_unref(outputFrame)
-        ret = av_buffersink_get_frame_flags(bufferSinkContext, outputFrame, 0)
-        guard ret == 0 else { return inputFrame }
-        return outputFrame ?? inputFrame
+        let ret = av_buffersrc_add_frame_flags(bufferSrcContext, inputFrame, 0)
+        if ret < 0 {
+            return
+        }
+        while av_buffersink_get_frame_flags(bufferSinkContext, inputFrame, 0) >= 0 {
+//                timebase = Timebase(av_buffersink_get_time_base(bufferSinkContext))
+            completionHandler(inputFrame)
+            // 一定要加av_frame_unref，不然会内存泄漏。
+            av_frame_unref(inputFrame)
+        }
     }
 }
