@@ -1,5 +1,5 @@
 //
-//  SubtitlePlayerItemTrack.swift
+//  SubtitleDecode.swift
 //  KSPlayer
 //
 //  Created by kintan on 2018/3/11.
@@ -14,16 +14,20 @@ import UIKit
 import AppKit
 #endif
 class SubtitleDecode: DecodeProtocol {
-    private let reg = try? NSRegularExpression(pattern: "\\{[^}]+\\}", options: .caseInsensitive)
     private var codecContext: UnsafeMutablePointer<AVCodecContext>?
-    private let scale = VideoSwresample(dstFormat: AV_PIX_FMT_ARGB)
+    private let scale = VideoSwresample(dstFormat: AV_PIX_FMT_ARGB, isDovi: false)
     private var subtitle = AVSubtitle()
     private var startTime = TimeInterval(0)
     private var preSubtitleFrame: SubtitleFrame?
+    private let assParse = AssParse()
     required init(assetTrack: FFmpegAssetTrack, options: KSOptions) {
-        startTime = assetTrack.startTime
+        startTime = assetTrack.startTime.seconds
         do {
-            codecContext = try assetTrack.ceateContext(options: options)
+            codecContext = try assetTrack.createContext(options: options)
+            if let pointer = codecContext?.pointee.subtitle_header {
+                let subtitleHeader = String(cString: pointer)
+                _ = assParse.canParse(scanner: Scanner(string: subtitleHeader))
+            }
         } catch {
             KSLog(error as CustomStringConvertible)
         }
@@ -42,30 +46,28 @@ class SubtitleDecode: DecodeProtocol {
         if gotsubtitle == 0 {
             return
         }
-        let (origin, attributedString, image) = text(subtitle: subtitle)
-        let position = packet.position
-        var start = packet.assetTrack.timebase.cmtime(for: position).seconds + TimeInterval(subtitle.start_display_time) / 1000.0
+        let timestamp = packet.timestamp
+        var start = packet.assetTrack.timebase.cmtime(for: timestamp).seconds + TimeInterval(subtitle.start_display_time) / 1000.0
         if start >= startTime {
             start -= startTime
         }
-        var duration = TimeInterval(subtitle.end_display_time - subtitle.start_display_time) / 1000.0
-        if duration == 0 {
+        var duration = 0.0
+        if subtitle.end_display_time != UInt32.max {
+            duration = TimeInterval(subtitle.end_display_time - subtitle.start_display_time) / 1000.0
+        }
+        if duration == 0, packet.duration != 0 {
             duration = packet.assetTrack.timebase.cmtime(for: packet.duration).seconds
         }
-        let part = SubtitlePart(start, start + duration, attributedString: attributedString)
-        part.image = image
-        part.origin = origin
-        let frame = SubtitleFrame(part: part, timebase: packet.assetTrack.timebase)
-        frame.position = position
-        if let preSubtitleFrame, preSubtitleFrame.part == part {
-            if let attributedString {
-                preSubtitleFrame.part.text?.append(NSAttributedString(string: "\n"))
-                preSubtitleFrame.part.text?.append(attributedString)
-            }
-        } else {
-            if let preSubtitleFrame, preSubtitleFrame.part.end == preSubtitleFrame.part.start {
-                preSubtitleFrame.part.end = frame.part.start
-            }
+        if let preSubtitleFrame, preSubtitleFrame.part.end == preSubtitleFrame.part.start {
+            preSubtitleFrame.part.end = start
+        }
+        preSubtitleFrame = nil
+        let parts = text(subtitle: subtitle)
+        for part in parts {
+            part.start = start
+            part.end = start + duration
+            let frame = SubtitleFrame(part: part, timebase: packet.assetTrack.timebase)
+            frame.timestamp = timestamp
             preSubtitleFrame = frame
             completionHandler(.success(frame))
         }
@@ -85,10 +87,11 @@ class SubtitleDecode: DecodeProtocol {
         }
     }
 
-    private func text(subtitle: AVSubtitle) -> (CGPoint, NSMutableAttributedString?, UIImage?) {
-        var attributedString: NSMutableAttributedString?
+    private func text(subtitle: AVSubtitle) -> [SubtitlePart] {
+        var parts = [SubtitlePart]()
         var images = [(CGRect, CGImage)]()
         var origin: CGPoint = .zero
+        var attributedString: NSMutableAttributedString?
         for i in 0 ..< Int(subtitle.num_rects) {
             guard let rect = subtitle.rects[i]?.pointee else {
                 continue
@@ -103,11 +106,8 @@ class SubtitleDecode: DecodeProtocol {
                 attributedString?.append(NSAttributedString(string: String(cString: text)))
             } else if let ass = rect.ass {
                 let scanner = Scanner(string: String(cString: ass))
-                if let group = AssParse.parse(scanner: scanner, reg: reg), let text = group.text {
-                    if attributedString == nil {
-                        attributedString = NSMutableAttributedString()
-                    }
-                    attributedString?.append(text)
+                if let group = assParse.parsePart(scanner: scanner) {
+                    parts.append(group)
                 }
             } else if rect.type == SUBTITLE_BITMAP {
                 if let image = scale.transfer(format: AV_PIX_FMT_PAL8, width: rect.w, height: rect.h, data: Array(tuple: rect.data), linesize: Array(tuple: rect.linesize))?.cgImage() {
@@ -115,14 +115,23 @@ class SubtitleDecode: DecodeProtocol {
                 }
             }
         }
-        if images.count > 1 {
-            origin = .zero
+        if images.count > 0 {
+            let part = SubtitlePart(0, 0, attributedString: nil)
+            if images.count > 1 {
+                origin = .zero
+            }
+            var image: UIImage?
+            // 因为字幕需要有透明度,所以不能用jpg；tif在iOS支持没有那么好，会有绿色背景； 用heic格式，展示的时候会卡主线程；所以最终用png。
+            if let data = CGImage.combine(images: images)?.data(type: .png, quality: 0.2) {
+                image = UIImage(data: data)
+            }
+            part.image = image
+            part.origin = origin
+            parts.append(part)
         }
-        var image: UIImage?
-        // 因为字幕需要有透明度,所以不能用jpg；tif在iOS支持没有那么好，会有绿色背景； 用heic格式，展示的时候会卡主线程；所以最终用png。
-        if let data = CGImage.combine(images: images)?.data(type: .png, quality: 0.2) {
-            image = UIImage(data: data)
+        if let attributedString {
+            parts.append(SubtitlePart(0, 0, attributedString: attributedString))
         }
-        return (origin, attributedString, image)
+        return parts
     }
 }
