@@ -11,12 +11,13 @@ import CoreGraphics
 import Foundation
 import SwiftUI
 
-public class SubtitlePart: CustomStringConvertible, NSMutableCopying {
-    public let start: TimeInterval
+public class SubtitlePart: CustomStringConvertible, Identifiable {
+    public var start: TimeInterval
     public var end: TimeInterval
     public var origin: CGPoint = .zero
-    public let text: NSMutableAttributedString?
+    public let text: NSAttributedString?
     public var image: UIImage?
+    public var textPosition: TextPosition?
     public var description: String {
         "Subtile Group ==========\nstart: \(start)\nend:\(end)\ntext:\(String(describing: text))"
     }
@@ -25,17 +26,70 @@ public class SubtitlePart: CustomStringConvertible, NSMutableCopying {
         var text = string
         text = text.trimmingCharacters(in: .whitespaces)
         text = text.replacingOccurrences(of: "\r", with: "")
-        self.init(start, end, attributedString: NSMutableAttributedString(string: text))
+        self.init(start, end, attributedString: NSAttributedString(string: text))
     }
 
-    public init(_ start: TimeInterval, _ end: TimeInterval, attributedString: NSMutableAttributedString?) {
+    public init(_ start: TimeInterval, _ end: TimeInterval, attributedString: NSAttributedString?) {
         self.start = start
         self.end = end
         text = attributedString
     }
+}
 
-    public func mutableCopy(with _: NSZone? = nil) -> Any {
-        SubtitlePart(start, end, attributedString: text?.mutableCopy() as? NSMutableAttributedString)
+public struct TextPosition {
+    public var verticalAlign: VerticalAlignment = .bottom
+    public var horizontalAlign: HorizontalAlignment = .center
+    public var leftMargin: CGFloat = 0
+    public var rightMargin: CGFloat = 0
+    public var verticalMargin: CGFloat = 10
+    public var edgeInsets: EdgeInsets {
+        var edgeInsets = EdgeInsets()
+        if verticalAlign == .bottom {
+            edgeInsets.bottom = verticalMargin
+        } else if verticalAlign == .top {
+            edgeInsets.top = verticalMargin
+        }
+        if horizontalAlign == .leading {
+            edgeInsets.leading = leftMargin
+        }
+        if horizontalAlign == .trailing {
+            edgeInsets.trailing = rightMargin
+        }
+        return edgeInsets
+    }
+
+    public mutating func ass(alignment: String?) {
+        switch alignment {
+        case "1":
+            verticalAlign = .bottom
+            horizontalAlign = .leading
+        case "2":
+            verticalAlign = .bottom
+            horizontalAlign = .center
+        case "3":
+            verticalAlign = .bottom
+            horizontalAlign = .trailing
+        case "4":
+            verticalAlign = .center
+            horizontalAlign = .leading
+        case "5":
+            verticalAlign = .center
+            horizontalAlign = .center
+        case "6":
+            verticalAlign = .center
+            horizontalAlign = .trailing
+        case "7":
+            verticalAlign = .top
+            horizontalAlign = .leading
+        case "8":
+            verticalAlign = .top
+            horizontalAlign = .center
+        case "9":
+            verticalAlign = .top
+            horizontalAlign = .trailing
+        default:
+            break
+        }
     }
 }
 
@@ -69,7 +123,7 @@ extension SubtitlePart: NumericComparable {
 }
 
 public protocol KSSubtitleProtocol {
-    func search(for time: TimeInterval) -> SubtitlePart?
+    func search(for time: TimeInterval) -> [SubtitlePart]
 }
 
 public protocol SubtitleInfo: KSSubtitleProtocol, AnyObject, Hashable, Identifiable {
@@ -79,7 +133,7 @@ public protocol SubtitleInfo: KSSubtitleProtocol, AnyObject, Hashable, Identifia
     //    var userInfo: NSMutableDictionary? { get set }
     //    var subtitleDataSouce: SubtitleDataSouce? { get set }
 //    var comment: String? { get }
-    func subtitle(isEnabled: Bool)
+    var isEnabled: Bool { get set }
 }
 
 public extension SubtitleInfo {
@@ -95,135 +149,55 @@ public extension SubtitleInfo {
 
 public class KSSubtitle {
     public var parts: [SubtitlePart] = []
-    public private(set) var currentIndex = 0 {
-        didSet {
-            if oldValue != currentIndex {
-                isChangeIndex = true
-            }
-        }
-    }
-
-    private var isFirstSearch = true
-    public var isChangeIndex = true
-    public var currentPart: SubtitlePart {
-        parts[currentIndex]
-    }
-
-    public var partsCount: Int {
-        parts.count
-    }
-
     public init() {}
 }
 
 extension KSSubtitle: KSSubtitleProtocol {
     /// Search for target group for time
-    public func search(for time: TimeInterval) -> SubtitlePart? {
-        var index = currentIndex
-        if searchIndex(for: time) != nil {
-            if currentIndex == index {
-                index = currentIndex + 1
-                if index < parts.count, parts[index] == time {
-                    // swiftlint:disable force_cast
-                    let copy = parts[currentIndex].mutableCopy() as! SubtitlePart
-                    // swiftlint:enable force_cast
-                    repeat {
-                        if let otherText = parts[index].text {
-                            copy.text?.append(NSAttributedString(string: "\n"))
-                            copy.text?.append(otherText)
-                        }
-                        copy.end = parts[index].end
-                        index += 1
-                    } while index < parts.count && parts[index] == time
-                    return copy
-                }
-            }
-            return parts[currentIndex]
-        } else {
-            return nil
-        }
-    }
-
-    public func searchIndex(for time: TimeInterval) -> Int? {
-        guard currentIndex < partsCount else {
-            return nil
-        }
-        let group = parts[currentIndex]
-        if group == time {
-            if isFirstSearch {
-                isChangeIndex = true
-                isFirstSearch = false
-            } else {
-                isChangeIndex = false
-            }
-            return currentIndex
-        } else if group < time, currentIndex + 1 < parts.count {
-            let group = parts[currentIndex + 1]
-            if group == time {
-                currentIndex += 1
-                return currentIndex
+    public func search(for time: TimeInterval) -> [SubtitlePart] {
+        var result = [SubtitlePart]()
+        for part in parts {
+            if part == time {
+                result.append(part)
+            } else if part.start > time {
+                break
             }
         }
-        if let firstIndex = parts.binarySearch(key: time) {
-            currentIndex = firstIndex
-            return currentIndex
-        }
-        return nil
-    }
-
-    public func searchIndex(filter: (SubtitlePart, Int) -> Bool) -> NSRange? {
-        var length = 0
-        for (index, group) in parts.enumerated() {
-            let count = group.text?.length ?? 0
-            if filter(group, length + count) {
-                if currentIndex != index {
-                    currentIndex = index
-                } else {
-                    if isFirstSearch {
-                        isChangeIndex = true
-                        isFirstSearch = false
-                    } else {
-                        isChangeIndex = false
-                    }
-                }
-                return NSRange(location: length, length: count)
-            } else {
-                length += count
-            }
-        }
-        return nil
+        return result
     }
 }
 
-public class KSURLSubtitle: KSSubtitle {
-    public func parse(url: URL, encoding: String.Encoding? = nil) async throws {
-        do {
-            var string: String?
-            let srtData = try await url.data()
-            let encodes = [encoding ?? String.Encoding.utf8,
-                           String.Encoding(rawValue: CFStringConvertEncodingToNSStringEncoding(CFStringEncoding(CFStringEncodings.big5.rawValue))),
-                           String.Encoding(rawValue: CFStringConvertEncodingToNSStringEncoding(CFStringEncoding(CFStringEncodings.GB_18030_2000.rawValue))),
-                           String.Encoding.unicode]
-            for encode in encodes {
-                string = String(data: srtData, encoding: encode)
-                if string != nil {
-                    break
-                }
+public extension KSSubtitle {
+    func parse(url: URL, userAgent: String? = nil, encoding: String.Encoding? = nil) async throws {
+        let data = try await url.data(userAgent: userAgent)
+        try parse(data: data, encoding: encoding)
+    }
+
+    func parse(data: Data, encoding: String.Encoding? = nil) throws {
+        var string: String?
+        let encodes = [encoding ?? String.Encoding.utf8,
+                       String.Encoding(rawValue: CFStringConvertEncodingToNSStringEncoding(CFStringEncoding(CFStringEncodings.big5.rawValue))),
+                       String.Encoding(rawValue: CFStringConvertEncodingToNSStringEncoding(CFStringEncoding(CFStringEncodings.GB_18030_2000.rawValue))),
+                       String.Encoding.unicode]
+        for encode in encodes {
+            string = String(data: data, encoding: encode)
+            if string != nil {
+                break
             }
-            guard let subtitle = string else {
-                throw NSError(errorCode: .subtitleUnEncoding, userInfo: ["url": url.absoluteString])
+        }
+        guard let subtitle = string else {
+            throw NSError(errorCode: .subtitleUnEncoding)
+        }
+        let scanner = Scanner(string: subtitle)
+        _ = scanner.scanCharacters(from: .controlCharacters)
+        let parse = KSOptions.subtitleParses.first { $0.canParse(scanner: scanner) }
+        if let parse {
+            parts = parse.parse(scanner: scanner)
+            if parts.count == 0 {
+                throw NSError(errorCode: .subtitleUnParse)
             }
-            let parse = KSOptions.subtitleParses.first { $0.canParse(subtitle: subtitle) }
-            if let parse {
-                parts = parse.parse(subtitle: subtitle)
-                if partsCount == 0 {
-                    throw NSError(errorCode: .subtitleUnParse, userInfo: ["url": url.absoluteString])
-                }
-            } else {
-                throw NSError(errorCode: .subtitleFormatUnSupport, userInfo: ["url": url.absoluteString])
-            }
-        } catch {
-            throw NSError(errorCode: .subtitleUnEncoding, userInfo: ["url": url.absoluteString])
+        } else {
+            throw NSError(errorCode: .subtitleFormatUnSupport)
         }
     }
 
@@ -264,24 +238,24 @@ open class SubtitleModel: ObservableObject {
         public var rawValue: CGFloat {
             switch self {
             case .smaller:
-                #if os(tvOS)
-                return 30
+                #if os(tvOS) || os(xrOS)
+                return 48
                 #elseif os(macOS)
                 return 20
                 #else
                 return 12
                 #endif
             case .standard:
-                #if os(tvOS)
-                return 36
+                #if os(tvOS) || os(xrOS)
+                return 58
                 #elseif os(macOS)
                 return 26
                 #else
                 return 16
                 #endif
             case .large:
-                #if os(tvOS)
-                return 42
+                #if os(tvOS) || os(xrOS)
+                return 68
                 #elseif os(macOS)
                 return 32
                 #else
@@ -300,95 +274,97 @@ open class SubtitleModel: ObservableObject {
     public static var textFontSize = SubtitleModel.Size.standard.rawValue
     public static var textBold = false
     public static var textItalic = false
-    public static var textXAlign: TextAlignment = .center
-    public static var textYAlign: VerticalAlignment = .bottom
-    public static var textXMargin: Int = 0
-    public static var textYMargin: Int = 10
-    public static var edgeInsets: EdgeInsets {
-        var edgeInsets = EdgeInsets()
-        if SubtitleModel.textYAlign == .bottom {
-            edgeInsets.bottom = CGFloat(SubtitleModel.textYMargin)
-        } else if SubtitleModel.textYAlign == .top {
-            edgeInsets.top = CGFloat(SubtitleModel.textYMargin)
-        }
-        if SubtitleModel.textXAlign == .leading {
-            edgeInsets.leading = CGFloat(SubtitleModel.textXMargin)
-        } else if SubtitleModel.textXAlign == .trailing {
-            edgeInsets.trailing = CGFloat(SubtitleModel.textXMargin)
-        }
-        return edgeInsets
-    }
-
+    public static var textPosition = TextPosition()
     private var subtitleDataSouces: [SubtitleDataSouce] = KSOptions.subtitleDataSouces
+    @Published
     public private(set) var subtitleInfos = [any SubtitleInfo]()
-    @Published public private(set) var part: SubtitlePart?
+    @Published
+    public private(set) var parts = [SubtitlePart]()
     public var subtitleDelay = 0.0 // s
     public var url: URL? {
         didSet {
             subtitleInfos.removeAll()
-            subtitleDataSouces.forEach { datasouce in
+            searchSubtitle(query: nil, languages: [])
+            for datasouce in subtitleDataSouces {
                 addSubtitle(dataSouce: datasouce)
+            }
+            // 要用async，不能在更新UI的时候，修改Publishe变量
+            DispatchQueue.main.async { [weak self] in
+                self?.parts = []
+                self?.selectedSubtitleInfo = nil
             }
         }
     }
 
-    @Published public var selectedSubtitleInfo: (any SubtitleInfo)? {
+    @Published
+    public var selectedSubtitleInfo: (any SubtitleInfo)? {
         didSet {
-            oldValue?.subtitle(isEnabled: false)
-            if let selectedSubtitleInfo {
-                addSubtitle(info: selectedSubtitleInfo)
-                selectedSubtitleInfo.subtitle(isEnabled: true)
+            oldValue?.isEnabled = false
+            selectedSubtitleInfo?.isEnabled = true
+            if let url, let info = selectedSubtitleInfo as? URLSubtitleInfo, !info.downloadURL.isFileURL, let cache = subtitleDataSouces.first(where: { $0 is CacheSubtitleDataSouce }) as? CacheSubtitleDataSouce {
+                cache.addCache(fileURL: url, downloadURL: info.downloadURL)
             }
         }
     }
 
     public init() {}
 
-    private func addSubtitle(info: any SubtitleInfo) {
+    public func addSubtitle(info: any SubtitleInfo) {
         if subtitleInfos.first(where: { $0.subtitleID == info.subtitleID }) == nil {
             subtitleInfos.append(info)
         }
     }
 
     public func subtitle(currentTime: TimeInterval) -> Bool {
-        let newPart: SubtitlePart?
+        var newParts = [SubtitlePart]()
         if let subtile = selectedSubtitleInfo {
-            let currentTime = currentTime - subtile.delay + subtitleDelay
-            if let part = subtile.search(for: currentTime) {
-                newPart = part
-            } else {
-                if let part, part.end > part.start, !(part == currentTime) {
-                    newPart = nil
-                } else {
-                    newPart = part
+            let currentTime = currentTime - subtile.delay - subtitleDelay
+            newParts = subtile.search(for: currentTime)
+            if newParts.isEmpty {
+                newParts = parts.filter { part in
+                    part.end <= part.start || part == currentTime
                 }
             }
-        } else {
-            newPart = nil
         }
         // swiftUI不会判断是否相等。所以需要这边判断下。
-        if newPart != part {
-            part = newPart
+        if newParts != parts {
+            for part in newParts {
+                if let text = part.text as? NSMutableAttributedString {
+                    text.addAttributes([.font: SubtitleModel.textFont],
+                                       range: NSRange(location: 0, length: text.length))
+                }
+            }
+            parts = newParts
             return true
         } else {
             return false
         }
     }
 
-    public func addSubtitle(dataSouce: SubtitleDataSouce) {
-        if let dataSouce = dataSouce as? SearchSubtitleDataSouce {
-            if let url {
-                Task {
-                    try? await dataSouce.searchSubtitle(url: url)
-                    dataSouce.infos.forEach { info in
-                        addSubtitle(info: info)
+    public func searchSubtitle(query: String?, languages: [String]) {
+        for dataSouce in subtitleDataSouces {
+            if let dataSouce = dataSouce as? SearchSubtitleDataSouce {
+                subtitleInfos.removeAll { info in
+                    dataSouce.infos.contains {
+                        $0 === info
                     }
                 }
+                Task { @MainActor in
+                    try? await dataSouce.searchSubtitle(query: query, languages: languages)
+                    subtitleInfos.append(contentsOf: dataSouce.infos)
+                }
+            }
+        }
+    }
+
+    public func addSubtitle(dataSouce: SubtitleDataSouce) {
+        if let dataSouce = dataSouce as? FileURLSubtitleDataSouce {
+            Task { @MainActor in
+                try? await dataSouce.searchSubtitle(fileURL: url)
+                subtitleInfos.append(contentsOf: dataSouce.infos)
             }
         } else {
-            dataSouce.infos.forEach { info in
-                addSubtitle(info: info)
-            }
+            subtitleInfos.append(contentsOf: dataSouce.infos)
         }
     }
 }
