@@ -1,5 +1,5 @@
 //
-//  SwiftUIView.swift
+//  AirPlayView.swift
 //  KSPlayer
 //
 //  Created by kintan on 2023/5/4.
@@ -7,6 +7,7 @@
 
 import AVKit
 import SwiftUI
+
 #if !os(xrOS)
 public struct AirPlayView: UIViewRepresentable {
     public init() {}
@@ -33,12 +34,44 @@ public struct AirPlayView: UIViewRepresentable {
 }
 #endif
 public extension View {
+    /// Applies the given transform if the given condition evaluates to `true`.
+    /// - Parameters:
+    ///   - condition: The condition to evaluate.
+    ///   - transform: The transform to apply to the source `View`.
+    /// - Returns: Either the original `View` or the modified `View` if the condition is `true`.
     @ViewBuilder
-    func `if`(_ condition: Bool, transform: (Self) -> some View) -> some View {
-        if condition {
+    func `if`(_ condition: @autoclosure () -> Bool, transform: (Self) -> some View) -> some View {
+        if condition() {
             transform(self)
         } else {
             self
         }
+    }
+
+    @ViewBuilder
+    func `if`(_ condition: @autoclosure () -> Bool, if ifTransform: (Self) -> some View, else elseTransform: (Self) -> some View) -> some View {
+        if condition() {
+            ifTransform(self)
+        } else {
+            elseTransform(self)
+        }
+    }
+
+    @ViewBuilder
+    func ifLet<T: Any>(_ optionalValue: T?, transform: (Self, T) -> some View) -> some View {
+        if let value = optionalValue {
+            transform(self, value)
+        } else {
+            self
+        }
+    }
+}
+
+extension Bool {
+    static var iOS16: Bool {
+        guard #available(iOS 16, *) else {
+            return true
+        }
+        return false
     }
 }
