@@ -10,6 +10,7 @@ import Combine
 import CoreServices
 import MediaPlayer
 import UIKit
+
 open class IOSVideoPlayerView: VideoPlayerView {
     private weak var originalSuperView: UIView?
     private var originalframeConstraints: [NSLayoutConstraint]?
@@ -33,6 +34,14 @@ open class IOSVideoPlayerView: VideoPlayerView {
             fullScreenDelegate?.player(isMaskShow: isMaskShow, isFullScreen: landscapeButton.isSelected)
         }
     }
+
+    #if !os(xrOS)
+    private var brightness: CGFloat = UIScreen.main.brightness {
+        didSet {
+            UIScreen.main.brightness = brightness
+        }
+    }
+    #endif
 
     override open func customizeUIComponents() {
         super.customizeUIComponents()
@@ -129,6 +138,9 @@ open class IOSVideoPlayerView: VideoPlayerView {
             }
             originalSuperView = superview
             originalframeConstraints = frameConstraints
+            if let originalframeConstraints {
+                NSLayoutConstraint.deactivate(originalframeConstraints)
+            }
             originalFrame = frame
             originalOrientations = viewController.supportedInterfaceOrientations
             let fullVC = PlayerFullScreenViewController(isHorizonal: isHorizonal)
@@ -152,17 +164,16 @@ open class IOSVideoPlayerView: VideoPlayerView {
                 return
             }
             let presentingVC = viewController.presentingViewController ?? viewController
-            KSOptions.supportedInterfaceOrientations = .portrait
+            if let originalOrientations {
+                KSOptions.supportedInterfaceOrientations = originalOrientations
+            }
             presentingVC.dismiss(animated: true) {
                 self.originalSuperView?.addSubview(self)
-                if let constraints = self.originalframeConstraints, constraints.count > 0 {
+                if let constraints = self.originalframeConstraints, !constraints.isEmpty {
                     NSLayoutConstraint.activate(constraints)
                 } else {
                     self.translatesAutoresizingMaskIntoConstraints = true
                     self.frame = self.originalFrame
-                }
-                if let originalOrientations = self.originalOrientations {
-                    KSOptions.supportedInterfaceOrientations = originalOrientations
                 }
             }
         }
@@ -177,11 +188,11 @@ open class IOSVideoPlayerView: VideoPlayerView {
             topMaskView.isHidden = KSOptions.topBarShowInCase != .always
         }
         toolBar.playbackRateButton.isHidden = false
-        toolBar.srtButton.isHidden = srtControl.subtitleInfos.count == 0
+        toolBar.srtButton.isHidden = srtControl.subtitleInfos.isEmpty
         if UIDevice.current.userInterfaceIdiom == .phone {
             if isLandscape {
                 landscapeButton.isHidden = true
-                toolBar.srtButton.isHidden = srtControl.subtitleInfos.count == 0
+                toolBar.srtButton.isHidden = srtControl.subtitleInfos.isEmpty
             } else {
                 toolBar.srtButton.isHidden = true
                 if let image = maskImageView.image {
@@ -252,7 +263,7 @@ open class IOSVideoPlayerView: VideoPlayerView {
                 }
             } else if KSOptions.enableBrightnessGestures {
                 #if !os(xrOS)
-                UIScreen.main.brightness += CGFloat(panValue(velocity: point, direction: direction, currentTime: Float(toolBar.currentTime), totalTime: Float(totalTime)))
+                brightness += CGFloat(panValue(velocity: point, direction: direction, currentTime: Float(toolBar.currentTime), totalTime: Float(totalTime)))
                 #endif
             }
         } else {
@@ -391,6 +402,7 @@ extension IOSVideoPlayerView: UIDocumentPickerDelegate {
 #endif
 
 #if os(iOS)
+@MainActor
 public class MenuController {
     public init(with builder: UIMenuBuilder) {
         builder.remove(menu: .format)
@@ -444,6 +456,7 @@ public class MenuController {
             NSLocalizedString("\(rawValue)", comment: "")
         }
 
+        @MainActor
         var command: String {
             switch self {
             case .rightArrow:
