@@ -8,12 +8,14 @@
 
 import AppKit
 import AVFoundation
+
 public extension NSPasteboard.PasteboardType {
     static let nsURL = NSPasteboard.PasteboardType("NSURL")
     static let nsFilenames = NSPasteboard.PasteboardType("NSFilenamesPboardType")
 }
 
 public extension NSDraggingInfo {
+    @MainActor
     func getUrl() -> URL? {
         guard let types = draggingPasteboard.types else { return nil }
 
@@ -43,8 +45,8 @@ open class MacVideoPlayerView: VideoPlayerView {
 
 extension MacVideoPlayerView {
     override open func updateTrackingAreas() {
-        trackingAreas.forEach {
-            removeTrackingArea($0)
+        for trackingArea in trackingAreas {
+            removeTrackingArea(trackingArea)
         }
         let trackingArea = NSTrackingArea(rect: bounds, options: [.mouseEnteredAndExited, .mouseMoved, .activeInKeyWindow], owner: self, userInfo: nil)
         addTrackingArea(trackingArea)
